### 2024-06-07 #6856 @editorialbot
[PRE REVIEW]: DataInterpolations.jl: Fast Interpolations of 1D data
**Submitting author:** @sathvikbhagavan (<a href="http://orcid.org/0000-0003-0785-3586">Sathvik Bhagavan</a>)
**Repository:** https://github.com/SciML/DataInterpolations.jl
**Branch with paper.md** (empty if default branch):
**Version:** v5.1.0
