### 2024-06-07 #6856 @danielskatz
@sathvikbhagavan - thanks for your submission.

Some issues I see:
* The license is called the MIT "Expat" license, but appears to just be the MIT licenses: https://opensource.org/license/mit Can you explain/discuss?
* Please work on the invalid and possibly missing DOIs that editorialbot suggests, but note that some may be incorrect. Please feel free to make changes to your .bib file, then use the command `@editorialbot check references` to check again, and the command `@editorialbot generate pdf` when the references are right to make a new PDF. editorialbot commands need to be the first entry in a new comment.
* The .md file needs a references header at the end, so that this will appear before the references in the generated PDF.  Please see the [example paper](https://joss.readthedocs.io/en/latest/example_paper.html)
