### 2024-06-07 #6856 @jbytecode
@sathvikbhagavan - Thank you for submitting your work to JOSS. I am the handling editor of this submission. Please firstly address the issues stressed by our track editor.

Secondly, I see many contributors including the ones with significant lines of code added in the repository. Why aren't they listed in the author list of the paper? Did you ask them to join writing the paper? Could you please clarify this issue?

Thank you in advance.
