### 2024-06-07 d68068e
refactor: integral function to correctly compute second index and remove `samples`
