### 2024-06-07 031a4b2
test: add more integral tests
