### 2024-06-11 86f09d7
docs: fix references for the JOSS paper
