### 2024-06-11 #6856 @ChrisRackauckas
> The license is called the MIT "Expat" license, but appears to just be the MIT licenses: https://opensource.org/license/mit Can you explain/discuss?

Calling it just the "MIT License" would be ambiguous which is why everyone refers to it as the MIT Expat License, originating from Expat. This is very standard and I do not know why something so standard would be flagged here?
