### 2024-06-11 #6856 @danielskatz
👋 @ChrisRackauckas - I note the wikipedia page is for the "MIT license", and it includes "The MIT License has the identifier MIT in the SPDX License List. It is also known as the Expat License." This says to me that the MIT license is the standard term.

JOSS requires an OSI-approved license, as defined in https://opensource.org/licenses-old/category.  Here, I see the "MIT license" defined. OSI doesn't list a license that includes expat in the name.

I accept that these are the same license, and that Expat is an alternative name in some cases, but it's not for OSI, so if would simply remove the first line of your LICENSE.md file, GitHub (and JOSS) would recognize this as the MIT license and our automated check would pass - see https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions/adding-a-license-to-a-repository for GitHub's recommendation for LICENSE files.
