### 2024-06-11 #6856 @ChrisRackauckas
Interesting, we just used the default MIT license generation from Gitkraken and that's how it gave it. Anyways, we can just make that look like the other standard. Odd that got flagged though since we've had other JOSS/JuliaCon Submissions (uses the same system) which didn't mention this.
