### 2024-06-13 9807717
refactor: add methods for computing with vector of timepoints for missing cases - vector of vectors and matrices
