### 2024-06-13 dfa2a97
test: add tests for computing with vector of timepoints
