### 2024-06-14 1bf37d0
docs: add References header in the paper
