+++ .typos.toml
@@ -1,1 +1,2 @@
 [default.extend-words]
+Commun = "Commun"
