### 2024-06-15 7dce110
docs: remove abstract of Schoenberg paper
