+++ joss/paper.bib
@@ -122,7 +122,6 @@
   publisher="Birkh{\"a}user Boston",
   address="Boston, MA",
   pages="3--57",
-  abstract="Introduction. Let there be given a sequence of ordinates<math display='block'><mrow><mrow><mo>{\{}</mo> <mrow><msub><mi>y</mi><mi>n</mi></msub></mrow> <mo>{\}}</mo></mrow><mtext>{\&}{\#}x2003;</mtext><mrow><mo>(</mo><mrow><mi>n</mi><mo>=</mo><mn>0</mn><mo>,</mo><mo>{\&}{\#}x00B1;</mo><mn>1</mn><mo>{\&}{\#}x00B1;</mo><mn>2</mn><mo>,</mo><mo>{\&}{\#}x2026;</mo></mrow><mo>)</mo></mrow><mo>,</mo></mrow></math>{\$}{\$}{\backslash}left{\backslash}{\{} {\{}{\{}y{\_}n{\}}{\}} {\backslash}right{\backslash}{\}}{\backslash}quad {\backslash}left( {\{}n = 0, {\backslash}pm 1 {\backslash}pm 2, {\backslash}ldots {\}} {\backslash}right),{\$}{\$}corresponding to all integral values of the variable x = n. If these ordinates are the values of a known analytic function F(x), then the problem of interpolation between these ordinates has an obvious and precise meaning: we are required to compute intermediate values F(x) to the same accuracy to which the ordinates are known. Undoubtedly, the most convenient tool for the solution of this problem is the polynomial central interpolation method. It uses the polynomial of degree k --- 1, interpolating k successive ordinates, as an approximation to F(x) only within a unit interval in x, centrally located with respect to its k defining ordinates. Assuming k fixed, successive approximating arcs for F(x) are thus obtained which present discontinuities on passing from one arc to the next if k is odd, or discontinuities in their first derivatives if k is even (see section 2.121). Actually these discontinuities are irrelevant in our present case of an analytic function F(x). Indeed, if the interpolated values obtained are sufficiently accurate, these discontinuities will be apparent only if we force the computation beyond the intrinsic accuracy of the yn.",
   isbn="978-1-4899-0433-1",
   doi="10.1007/978-1-4899-0433-1_1",
   url="https://doi.org/10.1007/978-1-4899-0433-1_1"
