### 2024-06-15 #6856 @ChrisRackauckas
> Secondly, I see many contributors including the ones with significant lines of code added in the repository. Why aren't they listed in the author list of the paper? Did you ask them to join writing the paper? Could you please clarify this issue?

I don't understand, who would you suggest? The next on the list had just a few trivial bug fixes and mostly just version bumps and CI changes. Then next on the list is Avik with one PR for adding derivative overloads: https://github.com/SciML/DataInterpolations.jl/pull/72 . I'll go with the ruling of whatever is the standard with JOSS, but my understanding was that it's not normal for folks who made one PR awhile back to be in the author list and it's mostly for the maintainers of the package?
