### 2024-06-15 #6856 @danielskatz
@ChrisRackauckas - please remember that questions and comments are generally well-intentioned. Regarding this case, it's really up to you to decide. Different software projects have different authorship policies, as reflected in different JOSS publications. Making your project's policy and decision process clear is what was being asked.
