### 2024-06-15 #6856 @jbytecode
@ChrisRackauckas, @sathvikbhagavan - Thank you for explaining why only a few of the participants in the project were designated as authors on the paper.

Now it is time to find reviewers for the process. Do you have any suggestions for the potential reviewers? You can use the search tool https://reviewers.joss.theoj.org/lookup to look up GitHub handles.

It seems the reviewers with Julia and math expertise is okay for this submission. Please mention their usernames without using the `@` character just because we don't want to bother them.

If you don't have any suggestions, I will find and assign two reviewers.

Thank you in advance.
