### 2024-06-15 #6856 @ChrisRackauckas
> please remember that questions and comments are generally well-intentioned

And indeed my questions are also well-intentioned. In the future it would be helpful for the questions to be more concrete so that they don't require clarifying questions to be understood (I'm still not entirely sure what you were asking for in a few cases here but it sounds like you're okay with the submission now so I guess all is well 🤷😅)
