### 2024-06-15 #6856 @jbytecode
The submission is okay, but I think the communication style used in this thread is a little bit tense. I don't think I can continue a process that could take months with this style of communication. May I withdraw from my editorial duties on this submission @danielskatz? I wanted to edit it especially because it was related to Julia. I am very sorry if I am causing you any trouble.
