### 2024-06-15 #6856 @ChrisRackauckas
I am sorry if anything is being misread and I agree @danielskatz made things a bit tense but I think everyone means well here. You haven't caused any trouble. I still don't entirely understand what was being asked, but it sounds like we're good.

For reviewers, I've worked with a good chunk of them so I'll try and make suggestions from the group that I haven't published with. matbesancon, crstnbr, dawbarton are folks who would have relevant background for which I don't think there's any COI to note.
