### 2024-06-18 #6856 @arfon
:wave: @matbesancon, @crstnbr, @dawbarton – would any of you be willing to review this submission for JOSS? The submission under consideration is " DataInterpolations.jl: Fast Interpolations of 1D data".

Many thanks
Arfon
