### 2024-06-22 #6917 @editorialbot
[REVIEW]: DataInterpolations.jl: Fast Interpolations of 1D data
**Submitting author:** @sathvikbhagavan (<a href="http://orcid.org/0000-0003-0785-3586">Sathvik Bhagavan</a>)
**Repository:** https://github.com/SciML/DataInterpolations.jl
**Branch with paper.md** (empty if default branch):
**Version:** v5.1.0
