### 2024-06-25 0c788e9
Introduce idx_prev to use as iguess, consistently use idx
