+++ src/interpolation_methods.jl
@@ -1,7 +1,9 @@
-function _interpolate(interp, t)
-    ((t < interp.t[1] || t > interp.t[end]) && !interp.extrapolate) &&
+function _interpolate(A, t)
+    ((t < A.t[1] || t > A.t[end]) && !A.extrapolate) &&
         throw(ExtrapolationError())
-    _interpolate(interp, t, firstindex(interp.t) - 1)[1]
+    val, idx_prev = _interpolate(A, t, A.idx_prev[])
+    A.idx_prev[] = idx_prev
+    return val
 end
 
 # Linear Interpolation
@@ -114,61 +116,62 @@ function _interpolate(A::LagrangeInterpolation{<:AbstractMatrix}, t::Number)
     N / D
 end
 
-function _interpolate(A::LagrangeInterpolation{<:AbstractVector}, t::Number, i)
-    _interpolate(A, t), i
+function _interpolate(A::LagrangeInterpolation{<:AbstractVector}, t::Number, idx)
+    _interpolate(A, t), idx
 end
 
-function _interpolate(A::LagrangeInterpolation{<:AbstractMatrix}, t::Number, i)
-    _interpolate(A, t), i
+function _interpolate(A::LagrangeInterpolation{<:AbstractMatrix}, t::Number, idx)
+    _interpolate(A, t), idx
 end
 
 function _interpolate(A::AkimaInterpolation{<:AbstractVector}, t::Number, iguess)
-    i = max(1, min(searchsortedlastcorrelated(A.t, t, iguess), length(A.t) - 1))
-    wj = t - A.t[i]
-    (@evalpoly wj A.u[i] A.b[i] A.c[i] A.d[i]), i
+    idx = max(1, min(searchsortedlastcorrelated(A.t, t, iguess), length(A.t) - 1))
+    wj = t - A.t[idx]
+    (@evalpoly wj A.u[idx] A.b[idx] A.c[idx] A.d[idx]), idx
 end
 
 # ConstantInterpolation Interpolation
 function _interpolate(A::ConstantInterpolation{<:AbstractVector}, t::Number, iguess)
     if A.dir === :left
         # :left means that value to the left is used for interpolation
-        i = max(1, searchsortedlastcorrelated(A.t, t, iguess))
-        return A.u[i], i
+        idx = max(1, searchsortedlastcorrelated(A.t, t, iguess))
+        return A.u[idx], idx
     else
         # :right means that value to the right is used for interpolation
-        i = min(length(A.t), searchsortedfirstcorrelated(A.t, t, iguess))
-        return A.u[i], i
+        idx = min(length(A.t), searchsortedfirstcorrelated(A.t, t, iguess))
+        return A.u[idx], idx
     end
 end
 
 function _interpolate(A::ConstantInterpolation{<:AbstractMatrix}, t::Number, iguess)
     if A.dir === :left
         # :left means that value to the left is used for interpolation
-        i = max(1, searchsortedlastcorrelated(A.t, t, iguess))
-        return A.u[:, i], i
+        idx = max(1, searchsortedlastcorrelated(A.t, t, iguess))
+        return A.u[:, idx], idx
     else
         # :right means that value to the right is used for interpolation
-        i = min(length(A.t), searchsortedfirstcorrelated(A.t, t, iguess))
[…]
