+++ src/interpolation_utils.jl
@@ -41,11 +41,12 @@ function munge_data(u::AbstractVector, t::AbstractVector)
     Tu = Base.nonmissingtype(eltype(u))
     Tt = Base.nonmissingtype(eltype(t))
     @assert length(t) == length(u)
-    non_missing_indices = collect(i
-    for i in 1:length(t)
-    if !ismissing(u[i]) && !ismissing(t[i]))
-    newu = Tu.([u[i] for i in non_missing_indices])
-    newt = Tt.([t[i] for i in non_missing_indices])
+    non_missing_indices = collect(
+        idx for idx in 1:length(t)
+        if !ismissing(u[idx]) && !ismissing(t[idx])
+    )
+    newu = Tu.([u[idx] for idx in non_missing_indices])
+    newt = Tt.([t[idx] for idx in non_missing_indices])
 
     return newu, newt
 end
@@ -54,11 +55,12 @@ function munge_data(U::StridedMatrix, t::AbstractVector)
     TU = Base.nonmissingtype(eltype(U))
     Tt = Base.nonmissingtype(eltype(t))
     @assert length(t) == size(U, 2)
-    non_missing_indices = collect(i
-    for i in 1:length(t)
-    if !any(ismissing, U[:, i]) && !ismissing(t[i]))
-    newUs = [TU.(U[:, i]) for i in non_missing_indices]
-    newt = Tt.([t[i] for i in non_missing_indices])
+    non_missing_indices = collect(
+        idx for idx in 1:length(t)
+        if !any(ismissing, U[:, idx]) && !ismissing(t[idx])
+    )
+    newUs = [TU.(U[:, idx]) for idx in non_missing_indices]
+    newt = Tt.([t[idx] for idx in non_missing_indices])
 
     return hcat(newUs...), newt
 end

[…]
