### 2024-06-26 054dd38
Use ReadOnlyArray for u, t of interpolations
