### 2024-06-26 6d9516e
Add safetycopy keyword
