+++ src/online.jl
@@ -18,22 +18,22 @@ function push!(di::ConstantInterpolation{U, T}, u::eltype(U), t::eltype(T)) wher
     di
 end
 
-function append!(di::LinearInterpolation{U, T}, u::U, t::T) where {U, T}
-    u, t = munge_data(u, t)
+function append!(di::LinearInterpolation{U, T}, u::U, t::T, safetycopy::Bool = false) where {U, T}
+    u, t = munge_data(u, t, safetycopy)
     append!(di.u, u)
     append!(di.t, t)
     di
 end
 
-function append!(di::QuadraticInterpolation{U, T}, u::U, t::T) where {U, T}
-    u, t = munge_data(u, t)
+function append!(di::QuadraticInterpolation{U, T}, u::U, t::T, safetycopy::Bool = false) where {U, T}
+    u, t = munge_data(u, t, safetycopy)
     append!(di.u, u)
     append!(di.t, t)
     di
 end
 
-function append!(di::ConstantInterpolation{U, T}, u::U, t::T) where {U, T}
-    u, t = munge_data(u, t)
+function append!(di::ConstantInterpolation{U, T}, u::U, t::T, safetycopy::Bool = false) where {U, T}
+    u, t = munge_data(u, t, safetycopy)
     append!(di.u, u)
     append!(di.t, t)
     di
