### 2024-06-26 c906e3b
Prototype of slope caching for LinearInterpolation
