+++ test/interpolation_tests.jl
@@ -1,6 +1,17 @@
 using DataInterpolations
+using FindFirstFunctions: searchsortedfirstcorrelated
 using StableRNGs
 using Optim, ForwardDiff
+
+function test_cached_index(A)
+    for t in range(first(A.t), last(A.t); length = 2 * length(A.t) - 1)
+        A(t)
+        if hasproperty(A, :idx_prev)
+            @test abs(A.idx_prev[] -
+                      searchsortedfirstcorrelated(A.t, t, A.idx_prev[])) <= 2
+        end
+    end
+end
 
 @testset "Linear Interpolation" begin
     for t in (1.0:10.0, 1.0collect(1:10))
@@ -45,6 +56,7 @@ using Optim, ForwardDiff
     @test A(0) == [-2.0 0.0; -3.0 0.0; -4.0 0.0]
     @test A(3) == [4.0 6.0; 6.0 9.0; 8.0 12.0]
     @test A(5) == [8.0 10.0; 12.0 15.0; 16.0 20.0]
+    test_cached_index(A)
 
     # with NaNs (#113)
     u = [NaN, 1.0, 2.0, 3.0]
@@ -168,6 +180,7 @@ end
     @test A(2.5) == 6.25
     @test A(3.5) == 12.25
     @test A(5.0) == 25
+    test_cached_index(A)
 
     # backward-looking interpolation
     u = [1.0, 4.0, 9.0, 16.0]
@@ -182,6 +195,7 @@ end
     @test A(2.5) == 6.25
     @test A(3.5) == 12.25
     @test A(5.0) == 25
+    test_cached_index(A)
 
     # Test both forward and backward-looking quadratic interpolation
     u = [1.0, 4.5, 6.0, 2.0]
@@ -203,6 +217,9 @@ end
     # In the last subinterval they should be the same again
     @test A_f(3.5) ≈ l₂ * u[2] + l₁ * u[3] + l₀ * u[4]
     @test A_b(3.5) ≈ l₂ * u[2] + l₁ * u[3] + l₀ * u[4]
+
+    test_cached_index(A_f)
+    test_cached_index(A_b)
 
     # Matrix interpolation test
     u = [1.0 4.0 9.0 16.0; 1.0 4.0 9.0 16.0]
@@ -321,6 +338,7 @@ end
     @test A(8.6) ≈ 4.1796554159017080820603951
     @test A(9.9) ≈ 3.4110386597938129327189927
     @test A(10.0) ≈ 3.0
+    test_cached_index(A)
 
     # Test extrapolation
     A = AkimaInterpolation(u, t; extrapolate = true)
@@ -345,6 +363,7 @@ end
         @test A(3.5) == u[1]
         @test A(4.0) == u[1]
         @test A(4.5) == u[1]
+        test_cached_index(A)
 
         A = ConstantInterpolation(u, t; extrapolate = true) # dir=:left is default
         @test A(0.5) == u[1]
@@ -356,6 +375,7 @@ end
         @test A(3.5) == u[3]
         @test A(4.0) == u[1]
         @test A(4.5) == u[1]
+        test_cached_index(A)
     end
 
     @testset "Matrix case" for u in [
@@ -372,6 +392,7 @@ end
         @test A(3.5) == u[:, 1]
         @test A(4.0) == u[:, 1]
         @test A(4.5) == u[:, 1]
+        test_cached_index(A)
 
         A = ConstantInterpolation(u, t; extrapolate = true) # dir=:left is default
         @test A(0.5) == u[:, 1]
@@ -383,6 +404,7 @@ end
         @test A(3.5) == u[:, 3]
         @test A(4.0) == u[:, 1]
         @test A(4.5) == u[:, 1]
+        test_cached_index(A)
     end
[…]
