### 2024-06-27 de154db
Support pushing/appending for LinearInterpolation
