+++ src/interpolation_methods.jl
@@ -14,7 +14,7 @@ function _interpolate(A::LinearInterpolation{<:AbstractVector}, t::Number, igues
         t1 = t2 = one(eltype(A.t))
         u1 = u2 = one(eltype(A.u))
     else
-        idx = max(1, min(searchsortedlastcorrelated(A.t, t, iguess), length(A.t) - 1))
+        idx = get_idx(A.t, t, iguess)
         t1, t2 = A.t[idx], A.t[idx + 1]
         u1, u2 = A.u[idx], A.u[idx + 1]
     end
@@ -27,7 +27,7 @@ function _interpolate(A::LinearInterpolation{<:AbstractVector}, t::Number, igues
 end
 
 function _interpolate(A::LinearInterpolation{<:AbstractMatrix}, t::Number, iguess)
-    idx = max(1, min(searchsortedlastcorrelated(A.t, t, iguess), length(A.t) - 1))
+    idx = get_idx(A.t, t, iguess)
     θ = (t - A.t[idx]) / (A.t[idx + 1] - A.t[idx])
     return (1 - θ) * A.u[:, idx] + θ * A.u[:, idx + 1], idx
 end
@@ -35,9 +35,7 @@ end
 # Quadratic Interpolation
 _quad_interp_indices(A, t) = _quad_interp_indices(A, t, firstindex(A.t) - 1)
 function _quad_interp_indices(A::QuadraticInterpolation, t::Number, iguess)
-    inner_idx = searchsortedlastcorrelated(A.t, t, iguess)
-    A.mode == :Backward && (inner_idx -= 1)
-    idx = max(1, min(inner_idx, length(A.t) - 2))
+    idx = get_idx(A.t, t, iguess; idx_shift = A.mode == :Backward ? -1 : 0, ub_shift = -2)
     idx, idx + 1, idx + 2
 end
 
@@ -125,7 +123,7 @@ function _interpolate(A::LagrangeInterpolation{<:AbstractMatrix}, t::Number, idx
 end
 
 function _interpolate(A::AkimaInterpolation{<:AbstractVector}, t::Number, iguess)
-    idx = max(1, min(searchsortedlastcorrelated(A.t, t, iguess), length(A.t) - 1))
+    idx = get_idx(A.t, t, iguess)
     wj = t - A.t[idx]
     (@evalpoly wj A.u[idx] A.b[idx] A.c[idx] A.d[idx]), idx
 end
@@ -134,30 +132,28 @@ end
 function _interpolate(A::ConstantInterpolation{<:AbstractVector}, t::Number, iguess)
     if A.dir === :left
         # :left means that value to the left is used for interpolation
-        idx = max(1, searchsortedlastcorrelated(A.t, t, iguess))
-        return A.u[idx], idx
+        idx = get_idx(A.t, t, iguess; lb = 1, ub_shift = 0)
     else
         # :right means that value to the right is used for interpolation
-        idx = min(length(A.t), searchsortedfirstcorrelated(A.t, t, iguess))
-        return A.u[idx], idx
+        idx = get_idx(A.t, t, iguess; side = :first, lb = 1, ub_shift = 0)
     end
+    A.u[idx], idx
 end
 
 function _interpolate(A::ConstantInterpolation{<:AbstractMatrix}, t::Number, iguess)
     if A.dir === :left
         # :left means that value to the left is used for interpolation
-        idx = max(1, searchsortedlastcorrelated(A.t, t, iguess))
-        return A.u[:, idx], idx
[…]
