### 2024-06-28 2b1af4d
Get LinearInterpolation derivative from cached slope
