### 2024-06-28 ef00740
Support push!/append! for ConstantInterpolation
