+++ src/online.jl
@@ -14,10 +14,10 @@ function push!(di::QuadraticInterpolation{U, T}, u::eltype(U), t::eltype(T)) whe
     di
 end
 
-function push!(di::ConstantInterpolation{U, T}, u::eltype(U), t::eltype(T)) where {U, T}
-    push!(di.u, u)
-    push!(di.t, t)
-    di
+function push!(A::ConstantInterpolation{U, T}, u::eltype(U), t::eltype(T)) where {U, T}
+    push!(A.u.parent, u)
+    push!(A.t.parent, t)
+    A
 end
 
 function append!(
@@ -34,11 +34,12 @@ function append!(
 end
 
 function append!(
-        di::QuadraticInterpolation{U, T}, u::U, t::T, safetycopy::Bool = false) where {U, T}
-    u, t = munge_data(u, t, safetycopy)
-    append!(di.u, u)
-    append!(di.t, t)
-    di
+    A::ConstantInterpolation{ReadOnlyVector{eltypeU, U}, ReadOnlyVector{eltypeT, T}}, u::U, t::T) where {
+        eltypeU, U, eltypeT, T}
+    u, t = munge_data(u, t, true)
+    append!(A.parent, u)
+    append!(A.parent.t, t)
+    A
 end
 
 function append!(
