### 2024-06-28 01fafaa
Add caching and online support for QuadraticInterpolation
