### 2024-06-28 e17f599
Compute integrals with cached parameters
