+++ src/DataInterpolations.jl
@@ -111,12 +111,13 @@ struct RegularizationSmooth{uType, tType, T, T2, ITP <: AbstractInterpolation{T}
             alg,
             Aitp,
             extrapolate)
-        new{typeof(u), typeof(t), eltype(u), typeof(λ), typeof(Aitp)}(u,
-            ReadOnlyArray(û),
-            t,
-            t̂,
-            ReadOnlyArray(wls),
-            ReadOnlyArray(wr),
+        new{typeof(u), typeof(t), eltype(u), typeof(λ), typeof(Aitp)}(
+            readonly_wrap(u),
+            readonly_wrap(û),
+            readonly_wrap(t),
+            readonly_wrap(t̂),
+            readonly_wrap(oftype(u.parent, wls)),
+            readonly_wrap(oftype(u.parent, wr)),
             d,
             λ,
             alg,
