### 2024-06-29 07b968c
Cache sigma for QuadraticSplineInterpolation
