+++ src/derivatives.jl
@@ -125,7 +125,7 @@ end
 # QuadraticSpline Interpolation
 function _derivative(A::QuadraticSpline{<:AbstractVector}, t::Number, iguess)
     idx = get_idx(A.t, t, iguess; lb = 2, ub_shift = 0, side = :first)
-    σ = 1 // 2 * (A.z[idx] - A.z[idx - 1]) / (A.t[idx] - A.t[idx - 1])
+    σ = A.p.σ[idx-1]
     A.z[idx - 1] + 2σ * (t - A.t[idx - 1]), idx
 end
 
