### 2024-06-29 1f9af50
Cache coefficients for CubicSplineInterpolation
