### 2024-06-29 282c840
Make non-forwarddiff calls of BSplineInterpolation, BSplineApprox non-allocating
