### 2024-06-29 c29c6ee
Non-allocating idx_prev using LagrangeInterpolation
