### 2024-06-29 61206b4
Merge pull request #264 from SciML/ChrisRackauckas-patch-2
Update Project.toml
