### 2024-06-29 5e39943
Only compute with nonzero spline coefficients
