### 2024-06-30 aa89a46
Add QuinticHermiteInterpolation
