+++ test/integral_tests.jl
@@ -113,7 +113,26 @@ end
     test_integral(AkimaInterpolation, u, t; name = "Akima Interpolation (Vector)")
     u = round.(rand(100), digits = 5)
     t = 1.0collect(1:100)
-    test_integral(AkimaInterpolation, u, t; name = "Akima Interpolation (Vector)")
+    test_integral(
+        AkimaInterpolation, u, t; name = "Akima Interpolation (Vector) with random points")
+end
+
+@testset "QuinticHermiteInterpolation" begin
+    u = [14.7, 11.51, 10.41, 14.95, 12.24, 11.22]
+    t = [0.0, 62.25, 109.66, 162.66, 205.8, 252.3]
+    du = [-0.047, -0.058, 0.054, 0.012, -0.068, 0.0]
+    ddu = [0.0, -0.00033, 0.0051, -0.0067, 0.0029, 0.0]
+    test_integral(QuinticHermiteInterpolation, u, t; args = [du, ddu],
+        name = "Quintic Hermite Interpolation (Vector)")
+
+    u = round.(rand(100), digits = 5)
+    t = 1.0collect(1:100)
+    du = diff(u) ./ diff(t)
+    push!(du, 0)
+    ddu = diff(du) ./ diff(t)
+    push!(ddu, 0)
+    test_integral(QuinticHermiteInterpolation, u, t; args = [du, ddu],
+        name = "Quintic Hermite Interpolation (Vector) with random points")
 end
 
 @testset "RegularizationSmooth" begin
