+++ src/integrals.jl
@@ -118,3 +118,16 @@ end
 function integral(A::BSplineApprox{<:AbstractVector{<:Number}}, t::Number)
     throw(IntegralNotFoundError())
 end
+
+# Quintic Hermite Interpolation
+function _integral(
+        A::QuinticHermiteInterpolation{<:AbstractVector{<:Number}}, idx::Number, t::Number)
+    Δt₀ = t - A.t[idx]
+    Δt₁ = t - A.t[idx + 1]
+    out = Δt₀ * (A.u[idx] + A.du[idx] * Δt₀ / 2 + A.ddu[idx] * Δt₀^2 / 6)
+    p = A.p.c₁[idx] + A.p.c₂[idx] * Δt₁ + A.p.c₃[idx] * Δt₁^2
+    dp = A.p.c₂[idx] + 2A.p.c₃[idx] * Δt₁
+    ddp = 2A.p.c₃[idx]
+    out += Δt₀^4 / 4 * (p - Δt₀ / 5 * dp + Δt₀^2 / 30 * ddp)
+    out
+end
