+++ src/interpolation_caches.jl
@@ -613,15 +613,50 @@ function BSplineApprox(u, t, d, h, pVecType, knotVecType; extrapolate = false)
 end
 
 """
-    QuinticHermiteInterpolation(u, t, du, ddu; extrapolate = false)
+    CubicHermiteInterpolation(du, u, t; extrapolate = false)
+
+It is a Cubic Hermite interpolation, which is a piece-wise third degree polynomial such that the value and the first derivative are equal to given values in the data points.
+
+## Arguments
+
+  - `du`: the derivative at the data points.
+  - `u`: data points.
+  - `t`: time points.
+
+## Keyword Arguments
+
+  - `extrapolate`: boolean value to allow extrapolation. Defaults to `false`.
+"""
+struct CubicHermiteInterpolation{uType, tType, duType, pType, T} <: AbstractInterpolation{T}
+    du::uType
+    u::uType
+    t::tType
+    p::CubicHermiteParameterCache{pType}
+    extrapolate::Bool
+    idx_prev::Base.RefValue{Int}
+    function CubicHermiteInterpolation(du, u, t, p, extrapolate)
+        new{typeof(u), typeof(t), typeof(du), typeof(p.c₁), eltype(u)}(
+            du, u, t, p, extrapolate, Ref(1))
+    end
+end
+
+function CubicHermiteInterpolation(du, u, t; extrapolate = false)
+    @assert length(u) == length(du)
+    u, t = munge_data(u, t)
+    p = CubicHermiteParameterCache(du, u, t)
+    return CubicHermiteInterpolation(du, u, t, p, extrapolate)
+end
+
+"""
+    QuinticHermiteInterpolation(ddu, du, u, t; extrapolate = false)
 
 It is a Quintic Hermite interpolation, which is a piece-wise fifth degree polynomial such that the value and the first and second derivative are equal to given values in the data points.
 
 ## Arguments
 
-  - `u`: data points.
-  - `du`: the derivative at the data points.
   - `ddu`: the second derivative at the data points.
+  - `du`: the derivative at the data points.
+  - `u`: data points.
   - `t`: time points.
 
 ## Keyword Arguments
@@ -630,22 +665,22 @@ It is a Quintic Hermite interpolation, which is a piece-wise fifth degree polyno
 """
 struct QuinticHermiteInterpolation{uType, tType, duType, dduType, pType, T} <:
        AbstractInterpolation{T}
+    ddu::uType
+    du::uType
     u::uType
     t::tType
-    du::duType
-    ddu::dduType
     p::QuinticHermiteParameterCache{pType}
     extrapolate::Bool
     idx_prev::Base.RefValue{Int}
-    function QuinticHermiteInterpolation(u, t, du, ddu, p, extrapolate)
+    function QuinticHermiteInterpolation(ddu, du, u, t, p, extrapolate)
         new{typeof(u), typeof(t), typeof(du), typeof(ddu), typeof(p.c₁), eltype(u)}(
-            u, t, du, ddu, p, extrapolate, Ref(1))
+            ddu, du, u, t, p, extrapolate, Ref(1))
     end
 end
 
-function QuinticHermiteInterpolation(u, t, du, ddu; extrapolate = false)
+function QuinticHermiteInterpolation(ddu, du, u, t; extrapolate = false)
     @assert length(u) == length(du) == length(ddu)
     u, t = munge_data(u, t)
-    p = QuinticHermiteParameterCache(u, t, du, ddu)
-    return QuinticHermiteInterpolation(u, t, du, ddu, p, extrapolate)
+    p = QuinticHermiteParameterCache(ddu, du, u, t)
+    return QuinticHermiteInterpolation(ddu, du, u, t, p, extrapolate)
 end
