+++ src/interpolation_methods.jl
@@ -202,6 +202,17 @@ function _interpolate(A::BSplineApprox{<:AbstractVector{<:Number}}, t::Number, i
     ucum, idx
 end
 
+# Cubic Hermite Interpolation
+function _interpolate(
+        A::CubicHermiteInterpolation{<:AbstractVector{<:Number}}, t::Number, iguess)
+    idx = get_idx(A.t, t, iguess)
+    Δt₀ = t - A.t[idx]
+    Δt₁ = t - A.t[idx + 1]
+    out = A.u[idx] + Δt₀ * A.du[idx]
+    out += Δt₀^2 * (A.p.c₁[idx] + Δt₁ * A.p.c₂[idx])
+    out, idx
+end
+
 # Quintic Hermite Interpolation
 function _interpolate(
         A::QuinticHermiteInterpolation{<:AbstractVector{<:Number}}, t::Number, iguess)
