+++ test/integral_tests.jl
@@ -5,8 +5,9 @@ using Optim, ForwardDiff
 using RegularizationTools
 using StableRNGs
 
-function test_integral(method, u, t; args = [], kwargs = [], name::String)
-    func = method(u, t, args...; kwargs..., extrapolate = true)
+function test_integral(method; args = [], kwargs = [], name::String)
+    func = method(args...; kwargs..., extrapolate = true)
+    (; t) = func
     t1 = minimum(t)
     t2 = maximum(t)
     @testset "$name" begin
@@ -46,7 +47,7 @@ function test_integral(method, u, t; args = [], kwargs = [], name::String)
         aint = integral(func, (t1 + t2) / 2, t2 + 5.0)
         @test isapprox(qint, aint, atol = 1e-6, rtol = 1e-8)
     end
-    func = method(u, t, args...; kwargs...)
+    func = method(args...; kwargs...)
     @test_throws DataInterpolations.ExtrapolationError integral(func, t[1] - 1.0)
     @test_throws DataInterpolations.ExtrapolationError integral(func, t[end] + 1.0)
     @test_throws DataInterpolations.ExtrapolationError integral(func, t[1] - 1.0, t[2])
@@ -56,27 +57,27 @@ end
 @testset "LinearInterpolation" begin
     u = 2.0collect(1:10)
     t = 1.0collect(1:10)
-    test_integral(LinearInterpolation, u, t; name = "Linear Interpolation (Vector)")
+    test_integral(
+        LinearInterpolation; args = [u, t], name = "Linear Interpolation (Vector)")
     u = round.(rand(100), digits = 5)
     t = 1.0collect(1:100)
-    test_integral(LinearInterpolation, u, t;
+    test_integral(LinearInterpolation; args = [u, t],
         name = "Linear Interpolation (Vector) with random points")
 end
 
 @testset "QuadraticInterpolation" begin
     u = [1.0, 4.0, 9.0, 16.0]
     t = [1.0, 2.0, 3.0, 4.0]
-    test_integral(QuadraticInterpolation, u, t; name = "Quadratic Interpolation (Vector)")
+    test_integral(
+        QuadraticInterpolation; args = [u, t], name = "Quadratic Interpolation (Vector)")
     u = [3.0, 0.0, 3.0, 0.0]
     t = [1.0, 2.0, 3.0, 4.0]
-    test_integral(QuadraticInterpolation,
-        u,
-        t;
-        args = [:Backward],
+    test_integral(QuadraticInterpolation;
+        args = [u, t, :Backward],
         name = "Quadratic Interpolation (Vector)")
     u = round.(rand(100), digits = 5)
     t = 1.0collect(1:10)
-    test_integral(QuadraticInterpolation, u, t;
+    test_integral(QuadraticInterpolation; args = [u, t],
         name = "Quadratic Interpolation (Vector) with random points")
 end
 
@@ -91,38 +92,53 @@ end
 @testset "QuadraticSpline" begin
     u = [0.0, 1.0, 3.0]
     t = [-1.0, 0.0, 1.0]
-    test_integral(QuadraticSpline, u, t; name = "Quadratic Spline (Vector)")
+    test_integral(QuadraticSpline; args = [u, t], name = "Quadratic Spline (Vector)")
     u = round.(rand(100), digits = 5)
     t = 1.0collect(1:100)
     test_integral(
-        QuadraticSpline, u, t; name = "Quadratic Spline (Vector) with random points")
[…]
