### 2024-07-03 fdce625
Let interpolations remember whether they did a safety copy
