+++ src/interpolation_caches.jl
@@ -20,15 +20,17 @@ struct LinearInterpolation{uType, tType, pType, T} <: AbstractInterpolation{T}
     p::LinearParameterCache{pType}
     extrapolate::Bool
     idx_prev::Base.RefValue{Int}
-    function LinearInterpolation(u, t, p, extrapolate)
-        new{typeof(u), typeof(t), typeof(p.slope), eltype(u)}(u, t, p, extrapolate, Ref(1))
+    safetycopy::Bool
+    function LinearInterpolation(u, t, p, extrapolate, safetycopy)
+        new{typeof(u), typeof(t), typeof(p.slope), eltype(u)}(
+            u, t, p, extrapolate, Ref(1), safetycopy)
     end
 end
 
 function LinearInterpolation(u, t; extrapolate = false, safetycopy = true)
     u, t = munge_data(u, t, safetycopy)
     p = LinearParameterCache(u, t)
-    LinearInterpolation(u, t, p, extrapolate)
+    LinearInterpolation(u, t, p, extrapolate, safetycopy)
 end
 
 """
@@ -55,24 +57,25 @@ struct QuadraticInterpolation{uType, tType, pType, T} <: AbstractInterpolation{T
     mode::Symbol
     extrapolate::Bool
     idx_prev::Base.RefValue{Int}
-    function QuadraticInterpolation(u, t, p, mode, extrapolate)
+    safetycopy::Bool
+    function QuadraticInterpolation(u, t, p, mode, extrapolate, safetycopy)
         mode ∈ (:Forward, :Backward) ||
             error("mode should be :Forward or :Backward for QuadraticInterpolation")
         new{typeof(u), typeof(t), typeof(p.l₀), eltype(u)}(
-            u, t, p, mode, extrapolate, Ref(1))
+            u, t, p, mode, extrapolate, Ref(1), safetycopy)
     end
 end
 
 function QuadraticInterpolation(u, t, mode; extrapolate = false, safetycopy = true)
     u, t = munge_data(u, t, safetycopy)
     p = QuadraticParameterCache(u, t)
-    QuadraticInterpolation(u, t, p, mode, extrapolate)
+    QuadraticInterpolation(u, t, p, mode, extrapolate, safetycopy)
 end
 
 function QuadraticInterpolation(u, t; extrapolate = false, safetycopy = true)
     u, t = munge_data(u, t, safetycopy)
     p = QuadraticParameterCache(u, t)
-    QuadraticInterpolation(u, t, p, :Forward, extrapolate)
+    QuadraticInterpolation(u, t, p, :Forward, extrapolate, safetycopy)
 end
 
 """
@@ -100,7 +103,8 @@ struct LagrangeInterpolation{uType, tType, T, bcacheType} <:
     idxs::Vector{Int}
     extrapolate::Bool
     idx_prev::Base.RefValue{Int}
-    function LagrangeInterpolation(u, t, n, extrapolate)
+    safetycopy::Bool
+    function LagrangeInterpolation(u, t, n, extrapolate, safetycopy)
         bcache = zeros(eltype(u[1]), n + 1)
         idxs = zeros(Int, n + 1)
         fill!(bcache, NaN)
@@ -110,7 +114,8 @@ struct LagrangeInterpolation{uType, tType, T, bcacheType} <:
             bcache,
             idxs,
             extrapolate,
-            Ref(1)
+            Ref(1),
+            safetycopy
         )
     end
 end
@@ -121,7 +126,7 @@ function LagrangeInterpolation(
     if n != length(t) - 1
         error("Currently only n=length(t) - 1 is supported")
     end
[…]
