### 2024-07-03 8410033
Test Interpolation types
