+++ test/interpolation_tests.jl
@@ -3,6 +3,18 @@ using FindFirstFunctions: searchsortedfirstcorrelated
 using StableRNGs
 using Optim, ForwardDiff
 
+function test_interpolation_type(T)
+    @test T <: DataInterpolations.AbstractInterpolation
+    @test hasfield(T, :u)
+    @test hasfield(T, :t)
+    @test hasfield(T, :extrapolate)
+    @test hasfield(T, :idx_prev)
+    @test hasfield(T, :safetycopy)
+    @test !isempty(methods(DataInterpolations._interpolate, (T, Any, Number)))
+    @test !isempty(methods(DataInterpolations._integral, (T, Any, Number)))
+    @test !isempty(methods(DataInterpolations._derivative, (T, Any, Number)))
+end
+
 function test_cached_index(A)
     for t in range(first(A.t), last(A.t); length = 2 * length(A.t) - 1)
         A(t)
@@ -13,6 +25,8 @@ function test_cached_index(A)
 end
 
 @testset "Linear Interpolation" begin
+    test_interpolation_type(LinearInterpolation)
+
     for t in (1.0:10.0, 1.0collect(1:10))
         u = 2.0collect(1:10)
         #t = 1.0collect(1:10)
@@ -159,6 +173,8 @@ end
 end
 
 @testset "Quadratic Interpolation" begin
+    test_interpolation_type(QuadraticInterpolation)
+
     u = [1.0, 4.0, 9.0, 16.0]
     t = [1.0, 2.0, 3.0, 4.0]
     A = QuadraticInterpolation(u, t; extrapolate = true)
@@ -254,6 +270,8 @@ end
 end
 
 @testset "Lagrange Interpolation" begin
+    test_interpolation_type(LagrangeInterpolation)
+
     u = [1.0, 4.0, 9.0]
     t = [1.0, 2.0, 3.0]
     A = LagrangeInterpolation(u, t)
@@ -312,6 +330,8 @@ end
 end
 
 @testset "Akima Interpolation" begin
+    test_interpolation_type(AkimaInterpolation)
+
     u = [0.0, 2.0, 1.0, 3.0, 2.0, 6.0, 5.5, 5.5, 2.7, 5.1, 3.0]
     t = collect(0.0:10.0)
     A = AkimaInterpolation(u, t)
@@ -341,6 +361,8 @@ end
 end
 
 @testset "ConstantInterpolation" begin
+    test_interpolation_type(ConstantInterpolation)
+
     t = [1.0, 2.0, 3.0, 4.0]
 
     @testset "Vector case" for u in [[1.0, 2.0, 0.0, 1.0], ["B", "C", "A", "B"]]
@@ -465,6 +487,8 @@ end
 end
 
 @testset "QuadraticSpline Interpolation" begin
+    test_interpolation_type(QuadraticSpline)
+
     u = [0.0, 1.0, 3.0]
     t = [-1.0, 0.0, 1.0]
 
@@ -510,6 +534,8 @@ end
 end
 
 @testset "CubicSpline Interpolation" begin
+    test_interpolation_type(CubicSpline)
+
     u = [0.0, 1.0, 3.0]
     t = [-1.0, 0.0, 1.0]
 
@@ -561,11 +587,13 @@ end
 end
 
 @testset "BSplines" begin
+
     # BSpline Interpolation and Approximation
     t = [0, 62.25, 109.66, 162.66, 205.8, 252.3]
     u = [14.7, 11.51, 10.41, 14.95, 12.24, 11.22]
 
     @testset "BSplineInterpolation" begin
+        test_interpolation_type(BSplineInterpolation)
[…]
