### 2024-07-03 7cb9ef8
Only allow adding data when `safetycopy = false`
