### 2024-07-04 a62531e
Cache cumulative integral
