+++ src/interpolation_caches.jl
@@ -26,7 +26,6 @@ end
 
 function LinearInterpolation(u, t; extrapolate = false)
     u, t = munge_data(u, t)
-    # TODO: Bad workaround, this computes the cached parameters twice
     A = LinearInterpolation(u, t, nothing, extrapolate)
     I = cumulative_integral(A)
     LinearInterpolation(u, t, I, extrapolate)
@@ -48,22 +47,25 @@ Extrapolation extends the last quadratic polynomial on each side.
 
   - `extrapolate`: boolean value to allow extrapolation. Defaults to `false`.
 """
-struct QuadraticInterpolation{uType, tType, T} <: AbstractInterpolation{T}
+struct QuadraticInterpolation{uType, tType, IType, T} <: AbstractInterpolation{T}
     u::uType
     t::tType
+    I::IType
     mode::Symbol
     extrapolate::Bool
     idx_prev::Base.RefValue{Int}
-    function QuadraticInterpolation(u, t, mode, extrapolate)
+    function QuadraticInterpolation(u, t, I, mode, extrapolate)
         mode ∈ (:Forward, :Backward) ||
             error("mode should be :Forward or :Backward for QuadraticInterpolation")
-        new{typeof(u), typeof(t), eltype(u)}(u, t, mode, extrapolate, Ref(1))
+        new{typeof(u), typeof(t), typeof(I), eltype(u)}(u, t, I, mode, extrapolate, Ref(1))
     end
 end
 
 function QuadraticInterpolation(u, t, mode; extrapolate = false)
     u, t = munge_data(u, t)
-    QuadraticInterpolation(u, t, mode, extrapolate)
+    A = QuadraticInterpolation(u, t, nothing, mode, extrapolate)
+    I = cumulative_integral(A)
+    QuadraticInterpolation(u, t, I, mode, extrapolate)
 end
 
 function QuadraticInterpolation(u, t; extrapolate = false)
@@ -129,19 +131,21 @@ Extrapolation extends the last cubic polynomial on each side.
 
   - `extrapolate`: boolean value to allow extrapolation. Defaults to `false`.
 """
-struct AkimaInterpolation{uType, tType, bType, cType, dType, T} <:
+struct AkimaInterpolation{uType, tType, IType, bType, cType, dType, T} <:
        AbstractInterpolation{T}
     u::uType
     t::tType
+    I::IType
     b::bType
     c::cType
     d::dType
     extrapolate::Bool
     idx_prev::Base.RefValue{Int}
-    function AkimaInterpolation(u, t, b, c, d, extrapolate)
-        new{typeof(u), typeof(t), typeof(b), typeof(c),
+    function AkimaInterpolation(u, t, I, b, c, d, extrapolate)
+        new{typeof(u), typeof(t), typeof(I), typeof(b), typeof(c),
             typeof(d), eltype(u)}(u,
             t,
+            I,
             b,
             c,
             d,
@@ -173,7 +177,9 @@ function AkimaInterpolation(u, t; extrapolate = false)
     c = (3.0 .* m[3:(end - 2)] .- 2.0 .* b[1:(end - 1)] .- b[2:end]) ./ dt
     d = (b[1:(end - 1)] .+ b[2:end] .- 2.0 .* m[3:(end - 2)]) ./ dt .^ 2
 
-    AkimaInterpolation(u, t, b, c, d, extrapolate)
+    A = AkimaInterpolation(u, t, nothing, b, c, d, extrapolate)
+    I = cumulative_integral(A)
+    AkimaInterpolation(u, t, I, b, c, d, extrapolate)
 end
 
 """
[…]
