### 2024-07-05 e568de7
Support inverting the integral of ConstantInterpolation
