+++ test/integral_inverse_tests.jl
@@ -21,3 +21,19 @@ end
     A = LinearInterpolation(u, t)
     @test_throws DataInterpolations.IntegralNotInvertibleError invert_integral(A)
 end
+
+@testset "Constant Interpolation" begin
+    t = collect(1:5)
+    u = [1.0, 1.0, 2.0, 4.0, 3.0]
+    test_integral_inverses(ConstantInterpolation; args = [u, t])
+    test_integral_inverses(ConstantInterpolation; args = [u, t], kwargs = [:dir => :right])
+
+    u = [1.0, -1.0, 2.0, 4.0, 3.0]
+    A = ConstantInterpolation(u, t)
+    @test_throws DataInterpolations.IntegralNotInvertibleError invert_integral(A)
+end
+
+t = collect(1:5)
+u = [1.0, 1.0, 2.0, 4.0, 3.0]
+A = QuadraticInterpolation(u, t)
+@test_throws DataInterpolations.IntegralInverseNotFoundError invert_integral(A)
