### 2024-07-05 946c47b
Invert when UnsafeAddDataError is called
