+++ src/online.jl
@@ -1,19 +1,17 @@
 import Base: append!, push!
 
 function push!(A::LinearInterpolation{U, T}, u::eltype(U), t::eltype(T)) where {U, T}
-    !A.safetycopy && throw(UnsafeAddDataError())
     push!(A.u.parent, u)
     push!(A.t.parent, t)
-    slope = LinearInterpolationParameters(A.u, A.t, length(A.t) - 1)
+    slope = linear_interpolation_parameters(A.u, A.t, length(A.t) - 1)
     push!(A.p.slope, slope)
     A
 end
 
 function push!(A::QuadraticInterpolation{U, T}, u::eltype(U), t::eltype(T)) where {U, T}
-    !A.safetycopy && throw(UnsafeAddDataError())
     push!(A.u.parent, u)
     push!(A.t.parent, t)
-    l₀, l₁, l₂ = QuadraticInterpolationParameters(A.u, A.t, length(A.t) - 2)
+    l₀, l₁, l₂ = quadratic_interpolation_parameters(A.u, A.t, length(A.t) - 2)
     push!(A.p.l₀, l₀)
     push!(A.p.l₁, l₁)
     push!(A.p.l₂, l₂)
@@ -21,7 +19,6 @@ function push!(A::QuadraticInterpolation{U, T}, u::eltype(U), t::eltype(T)) wher
 end
 
 function push!(A::ConstantInterpolation{U, T}, u::eltype(U), t::eltype(T)) where {U, T}
-    !A.safetycopy && throw(UnsafeAddDataError())
     push!(A.u.parent, u)
     push!(A.t.parent, t)
     A
@@ -30,12 +27,11 @@ end
 function append!(
         A::LinearInterpolation{ReadOnlyVector{eltypeU, U}, ReadOnlyVector{eltypeT, T}}, u::U, t::T) where {
         eltypeU, U, eltypeT, T}
-    !A.safetycopy && throw(UnsafeAddDataError())
     length_old = length(A.t)
     u, t = munge_data(u, t, true)
     append!(A.u.parent, u)
     append!(A.t.parent, t)
-    slope = LinearInterpolationParameters.(
+    slope = linear_interpolation_parameters.(
         Ref(A.u), Ref(A.t), length_old:(length(A.t) - 1))
     append!(A.p.slope, slope)
     A
@@ -44,7 +40,6 @@ end
 function append!(
         A::ConstantInterpolation{ReadOnlyVector{eltypeU, U}, ReadOnlyVector{eltypeT, T}}, u::U, t::T) where {
         eltypeU, U, eltypeT, T}
-    !A.safetycopy && throw(UnsafeAddDataError())
     u, t = munge_data(u, t, true)
     append!(A.u.parent, u)
     append!(A.t.parent, t)
@@ -54,12 +49,11 @@ end
 function append!(
         A::QuadraticInterpolation{ReadOnlyVector{eltypeU, U}, ReadOnlyVector{eltypeT, T}}, u::U, t::T) where {
         eltypeU, U, eltypeT, T}
-    !A.safetycopy && throw(UnsafeAddDataError())
     length_old = length(A.t)
     u, t = munge_data(u, t, true)
     append!(A.u.parent, u)
     append!(A.t.parent, t)
-    parameters = QuadraticInterpolationParameters.(
+    parameters = quadratic_interpolation_parameters.(
         Ref(A.u), Ref(A.t), (length_old - 1):(length(A.t) - 2))
     l₀, l₁, l₂ = collect.(eachrow(hcat(collect.(parameters)...)))
     append!(A.p.l₀, l₀)
