+++ src/DataInterpolations.jl
@@ -81,12 +81,6 @@ function Base.showerror(io::IO, e::MustCopyError)
     print(io, MUST_COPY_ERROR)
 end
 
-const UNSAFE_ADD_DATA_ERROR = "Can only add data to the interpolation if safetycopy = true."
-struct UnsafeAddDataError <: Exception end
-function Base.showerror(io::IO, e::UnsafeAddDataError)
-    print(io, UNSAFE_ADD_DATA_ERROR)
-end
-
 export LinearInterpolation, QuadraticInterpolation, LagrangeInterpolation,
        AkimaInterpolation, ConstantInterpolation, QuadraticSpline, CubicSpline,
        BSplineInterpolation, BSplineApprox, CubicHermiteSpline,
