+++ test/online_tests.jl
@@ -1,10 +1,10 @@
 using DataInterpolations
 
-t1 = Float64[1, 2, 3]
-u1 = Float64[0, 1, 0]
+t1 = [1.0, 2.0, 3.0]
+u1 = [0.0, 1.0, 0.0]
 
-t2 = Float64[4, 5, 6]
-u2 = Float64[1, 2, 1]
+t2 = [4.0, 5.0, 6.0]
+u2 = Float64[1.0, 2.0, 1.0]
 
 ts = 1.0:0.5:6.0
 

[…]
