### 2024-07-05 3a05b0a
Add name and affiliation to paper
