+++ test/online_tests.jl
@@ -4,7 +4,7 @@ t1 = [1.0, 2.0, 3.0]
 u1 = [0.0, 1.0, 0.0]
 
 t2 = [4.0, 5.0, 6.0]
-u2 = Float64[1.0, 2.0, 1.0]
+u2 = [1.0, 2.0, 1.0]
 
 ts = 1.0:0.5:6.0
 
