+++ src/integral_inverses.jl
@@ -1,19 +1,25 @@
 abstract type AbstractIntegralInverseInterpolation{T} <: AbstractInterpolation{T} end
 
-_integral(A::AbstractIntegralInverseInterpolation, idx, t) = throw(IntegralNotFoundError())
-
 """
     invert_integral(A::AbstractInterpolation)::AbstractIntegralInverseInterpolation
 
 Creates the inverted integral interpolation object from the given interpolation. Conditions:
+
   - The range of `A` must be strictly positive
   - There must be an ordering defined on the data type of `A.u`
   - This is currently only supported for ConstantInterpolation and LinearInterpolation
 
 ## Arguments
+
   - `A`: interpolation object satisfying the above requirements
 """
 invert_integral(A::AbstractInterpolation) = throw(IntegralInverseNotFoundError())
+
+_integral(A::AbstractIntegralInverseInterpolation, idx, t) = throw(IntegralNotFoundError())
+
+function _derivative(A::AbstractIntegralInverseInterpolation, t::Number, iguess)
+    inv(A.itp(A(t))), A.idx_prev[]
+end
 
 """
     some stuff
