### 2024-07-06 13fcf22
docs: update series types recipes available
