### 2024-07-06 6197238
docs: add mention of Hermite Splines in joss paper
