+++ test/online_tests.jl
@@ -1,18 +1,31 @@
 using DataInterpolations
 
-t = [1, 2, 3]
-u = [0, 1, 0]
+t1 = [1.0, 2.0, 3.0]
+u1 = [0.0, 1.0, 0.0]
 
-for di in [LinearInterpolation, QuadraticInterpolation, ConstantInterpolation]
-    li = di(copy(u), copy(t))
-    append!(li, u, t)
-    li2 = di(vcat(u, u), vcat(t, t))
-    @test li.u == li2.u
-    @test li.t == li2.t
+t2 = [4.0, 5.0, 6.0]
+u2 = [1.0, 2.0, 1.0]
 
-    li = di(copy(u), copy(t))
-    push!(li, 1, 4)
-    li2 = di(vcat(u, 1), vcat(t, 4))
-    @test li.u == li2.u
-    @test li.t == li2.t
+ts = 1.0:0.5:6.0
+
+for method in [LinearInterpolation, QuadraticInterpolation, ConstantInterpolation]
+    func1 = method(u1, t1)
+    append!(func1, u2, t2)
+    func2 = method(vcat(u1, u2), vcat(t1, t2))
+    @test func1.u == func2.u
+    @test func1.t == func2.t
+    for name in propertynames(func1.p)
+        @test getfield(func1.p, name) == getfield(func2.p, name)
+    end
+    @test func1(ts) == func2(ts)
+
+    func1 = method(u1, t1)
+    push!(func1, 1.0, 4.0)
+    func2 = method(vcat(u1, 1.0), vcat(t1, 4.0))
+    @test func1.u == func2.u
+    @test func1.t == func2.t
+    for name in propertynames(func1.p)
+        @test getfield(func1.p, name) == getfield(func2.p, name)
+    end
+    @test func1(ts) == func2(ts)
 end
