+++ src/online.jl
@@ -1,40 +1,63 @@
 import Base: append!, push!
 
-function push!(di::LinearInterpolation{U, T}, u::eltype(U), t::eltype(T)) where {U, T}
-    push!(di.u, u)
-    push!(di.t, t)
-    di
+function push!(A::LinearInterpolation{U, T}, u::eltype(U), t::eltype(T)) where {U, T}
+    push!(A.u.parent, u)
+    push!(A.t.parent, t)
+    slope = linear_interpolation_parameters(A.u, A.t, length(A.t) - 1)
+    push!(A.p.slope, slope)
+    A
 end
 
-function push!(di::QuadraticInterpolation{U, T}, u::eltype(U), t::eltype(T)) where {U, T}
-    push!(di.u, u)
-    push!(di.t, t)
-    di
+function push!(A::QuadraticInterpolation{U, T}, u::eltype(U), t::eltype(T)) where {U, T}
+    push!(A.u.parent, u)
+    push!(A.t.parent, t)
+    l₀, l₁, l₂ = quadratic_interpolation_parameters(A.u, A.t, length(A.t) - 2)
+    push!(A.p.l₀, l₀)
+    push!(A.p.l₁, l₁)
+    push!(A.p.l₂, l₂)
+    A
 end
 
-function push!(di::ConstantInterpolation{U, T}, u::eltype(U), t::eltype(T)) where {U, T}
-    push!(di.u, u)
-    push!(di.t, t)
-    di
+function push!(A::ConstantInterpolation{U, T}, u::eltype(U), t::eltype(T)) where {U, T}
+    push!(A.u.parent, u)
+    push!(A.t.parent, t)
+    A
 end
 
-function append!(di::LinearInterpolation{U, T}, u::U, t::T) where {U, T}
-    u, t = munge_data(u, t)
-    append!(di.u, u)
-    append!(di.t, t)
-    di
+function append!(
+        A::LinearInterpolation{ReadOnlyVector{eltypeU, U}, ReadOnlyVector{eltypeT, T}}, u::U, t::T) where {
+        eltypeU, U, eltypeT, T}
+    length_old = length(A.t)
+    u, t = munge_data(u, t, true)
+    append!(A.u.parent, u)
+    append!(A.t.parent, t)
+    slope = linear_interpolation_parameters.(
+        Ref(A.u), Ref(A.t), length_old:(length(A.t) - 1))
+    append!(A.p.slope, slope)
+    A
 end
 
-function append!(di::QuadraticInterpolation{U, T}, u::U, t::T) where {U, T}
-    u, t = munge_data(u, t)
-    append!(di.u, u)
-    append!(di.t, t)
-    di
+function append!(
+        A::ConstantInterpolation{ReadOnlyVector{eltypeU, U}, ReadOnlyVector{eltypeT, T}}, u::U, t::T) where {
+        eltypeU, U, eltypeT, T}
+    u, t = munge_data(u, t, true)
+    append!(A.u.parent, u)
+    append!(A.t.parent, t)
+    A
 end
 
-function append!(di::ConstantInterpolation{U, T}, u::U, t::T) where {U, T}
-    u, t = munge_data(u, t)
-    append!(di.u, u)
-    append!(di.t, t)
-    di
+function append!(
+        A::QuadraticInterpolation{ReadOnlyVector{eltypeU, U}, ReadOnlyVector{eltypeT, T}}, u::U, t::T) where {
+        eltypeU, U, eltypeT, T}
+    length_old = length(A.t)
+    u, t = munge_data(u, t, true)
+    append!(A.u.parent, u)
+    append!(A.t.parent, t)
[…]

[…]
