+++ dev/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"interface/#Interface-for-using-the-Interpolations-object","page":"Interface","title":"Interface for using the Interpolations object","text":"","category":"section"},{"location":"interface/","page":"Interface","title":"Interface","text":"We will again use the same data as the previous tutorial to demonstrate how to use the Interpolations object for computing interpolated values at any time point, as well as derivatives and integrals.","category":"page"},{"location":"interface/","page":"Interface","title":"Interface","text":"using DataInterpolations\n\n# Dependent variable\nu = [14.7, 11.51, 10.41, 14.95, 12.24, 11.22]\n\n# Independent variable\nt = [0.0, 62.25, 109.66, 162.66, 205.8, 252.3]","category":"page"},{"location":"interface/#Interpolated-values","page":"Interface","title":"Interpolated values","text":"","category":"section"},{"location":"interface/","page":"Interface","title":"Interface","text":"All interpolation methods return an object from which we can compute the value of the dependent variable at any time point.","category":"page"},{"location":"interface/","page":"Interface","title":"Interface","text":"We will use the CubicSpline method for demonstration, but the API is the same for all the methods. We can also pass the extrapolate = true keyword if we want to allow the interpolation to go beyond the range of the timepoints. The default value is extrapolate = false.","category":"page"},{"location":"interface/","page":"Interface","title":"Interface","text":"A1 = CubicSpline(u, t)\n\n# For interpolation do, A(t)\nA1(100.0)\n\nA2 = CubicSpline(u, t; extrapolate = true)\n\n# Extrapolation\nA2(300.0)","category":"page"},{"location":"interface/","page":"Interface","title":"Interface","text":"note: Note\nThe values computed beyond the range of the time points provided during interpolation will not be reliable, as these methods only perform well within the range and the first/last piece polynomial fit is extrapolated on either side which might not reflect the true nature of the data.","category":"page"},{"location":"interface/","page":"Interface","title":"Interface","text":"The keyword safetycopy = false can be passed to make sure no copies of u and t are made when initializing the interpolation object.","category":"page"},{"location":"interface/","page":"Interface","title":"Interface","text":"A3 = QuadraticInterpolation(u, t; safetycopy = false)\n\n# Check for same memory\nu === A3.u.parent","category":"page"},{"location":"interface/","page":"Interface","title":"Interface","text":"Note that this does not prevent allocation in every interpolation constructor call, because parameter values are cached for all interpolation types except ConstantInterpolation.","category":"page"},{"location":"interface/","page":"Interface","title":"Interface","text":"Because of the caching of parameters which depend on u and t, this data should not be mutated. Therefore u and t are wrapped in a ReadOnlyArray from ReadOnlyArrays.jl.","category":"page"},{"location":"interface/","page":"Interface","title":"Interface","text":"A3.t[2] = 3.14","category":"page"},{"location":"interface/#Derivatives","page":"Interface","title":"Derivatives","text":"","category":"section"},{"location":"interface/","page":"Interface","title":"Interface","text":"Derivatives of the interpolated curves can also be computed at any point for all the methods. Derivatives upto second order is supported where first order derivative is computed analytically and second order using ForwardDiff.jl. Order is passed as the third argument. It is 1 by default.","category":"page"},{"location":"interface/","page":"Interface","title":"Interface","text":"We will continue with the ab
[…]
