### 2024-07-06 a4c2988
Update integral values when adding data
