### 2024-07-06 f3350bb
refactor: use BaryCentric Formula for Lagrange Interpolation
