### 2024-07-06 c17fc26
refactor: derivatives with BaryCentric Formula for Lagrange Interpolation
