### 2024-07-06 d5786ea
test: approx for an exterior point instead of exact equality test
