### 2024-07-07 2fe49b3
refactor: error in BSpline when number of data/control is less than or equal to degree
