+++ src/interpolation_caches.jl
@@ -481,6 +481,7 @@ function BSplineInterpolation(
         u, t, d, pVecType, knotVecType; extrapolate = false, safetycopy = true)
     u, t = munge_data(u, t, safetycopy)
     n = length(t)
+    n < d + 1 && error("BSplineInterpolation needs at least d + 1, i.e. $(d+1) points.")
     s = zero(eltype(u))
     p = zero(t)
     k = zeros(eltype(t), n + d + 1)
@@ -615,6 +616,7 @@ function BSplineApprox(
         u, t, d, h, pVecType, knotVecType; extrapolate = false, safetycopy = true)
     u, t = munge_data(u, t, safetycopy)
     n = length(t)
+    h < d + 1 && error("BSplineApprox needs at least d + 1, i.e. $(d+1) control points.")
     s = zero(eltype(u))
     p = zero(t)
     k = zeros(eltype(t), h + d + 1)
