### 2024-07-07 19b1db9
test: add tests for checking error with bspline wrt to degree and data/control points
