### 2024-07-07 ab43395
fixup! refactor: use BaryCentric Formula for Lagrange Interpolation
