### 2024-07-07 9737c50
half a LinearInterpolation POC
