### 2024-07-07 16e985e
Add computations in POC
