+++ ext/DataInterpolationsChainRulesCoreExt.jl
@@ -14,38 +14,55 @@ else
     using ..ChainRulesCore
 end
 
-function ChainRulesCore.rrule(::typeof(linear_interpolation_parameters), u, t, idx)
+function ChainRulesCore.rrule(::typeof(linear_interpolation_parameters), u::AbstractVector, t::AbstractVector, idx::Integer)
     slope = linear_interpolation_parameters(u, t, idx)
+    Δt = t[idx + 1] - t[idx]
+    Δu = u[idx + 1] - t[idx]
+    # TODO: use sparse arrays
+    du = zero(u)
+    dt = zero(t)
 
-    function linear_interpolation_parameters_pullback(Δslope)
-        du = @thunk(Δslope*...) # TODO: How to handle sparsity?
-        dt = @thunk(Δslope*...) # TODO: How to handle sparsity?
+    function linear_interpolation_parameters_pullback(Δ)
         df = NoTangent()
+        du .= zero(eltype(u))
+        du[idx] = - Δ / Δt
+        du[idx + 1] = Δslope / Δt
+        dt .= zero(eltype(t))
+        dt[idx] = Δ * Δu / Δt^2
+        dt[idx + 1] = - Δ * Δu / Δt^2
         didx = NoTangent()
-
         return (df, du, dt, didx)
     end
 
     return slope, linear_interpolation_parameters_pullback
 end
 
-function _tangent_u(A::LinearInterpolation, t)
-    ... # TODO: How to handle sparsity?
+function _tangent_u!(Δu::AbstractVector, A::LinearInterpolation)
+    Δu .= zero(eltype(A.u))
+    Δu
 end
 
-function _tangent_t(A::LinearInterpolation, t)
-    ... # TODO: How to handle sparsity?
+function _tangent_t!(Δt::AbstractVector, A::LinearInterpolation)
+    idx = A.idx_prev[]
+    Δt .= zero(eltype(Δt))
+    Δt[idx] = one(eltype(Δt))
+    Δt
 end
 
 function ChainRulesCore.rrule(::typeof(_interpolate), A::AType, t, iguess) where {AType}
-    u = _interpolate(A, t, iguess)[1]
+    u = _interpolate(A, t)
+    # TODO: use sparse arrays
+    Δu = zero(A.u)
+    Δt = zero(A.t)
+
     function _interpolate_pullback(Δ)
         df = NoTangent()
-        dA = Tangent{AType}(; u = _tangent_u(A, t), t = _tangent_t(A, t))
+        dA = Tangent{AType}(; u = _tangent_u!(Δu, A), t = _tangent_t!(Δt, A))
         dt = @thunk(derivative(A, t)*Δ)
         diguess = NoTangent()
         return df, dA, dt, diguess
     end
+
     u, _interpolate_pullback
 end
 
