+++ src/parameter_caches.jl
@@ -5,14 +5,12 @@ end
 function LinearParameterCache(u, t)
     slope_prototype = linear_interpolation_parameters(u, t, 1)
     idxs = 1:(length(t) - 1)
-    slope = [zero(slope_prototype) for i in idxs]
-    for idx in idxs
-        slope[idx] = linear_interpolation_parameters(u, t, idx)
-    end
+    slope = [linear_interpolation_parameters(u, t, idx) for idx in idxs]
     return LinearParameterCache(slope)
 end
 
-function linear_interpolation_parameters(u::AbstractVector, t::AbstractVector, idx::Integer)::Number
+function linear_interpolation_parameters(
+        u::AbstractVector, t::AbstractVector, idx::Integer)::Number
     Δu = u isa AbstractMatrix ? u[:, idx + 1] - u[:, idx] : u[idx + 1] - u[idx]
     Δt = t[idx + 1] - t[idx]
     slope = Δu / Δt
