### 2024-07-08 e67abcd
I now get gradients w.r.t. the u of the datapoints, but they are probably not correct yet
