### 2024-07-08 b1b5004
gradients w.r.t. u PoC!
