+++ ext/DataInterpolationsChainRulesCoreExt.jl
@@ -4,33 +4,31 @@ if isdefined(Base, :get_extension)
     using DataInterpolations: _interpolate, derivative, AbstractInterpolation,
                               LagrangeInterpolation, AkimaInterpolation,
                               BSplineInterpolation, BSplineApprox, LinearInterpolation,
-                              linear_interpolation_parameters
+                              linear_interpolation_parameters, get_idx
     using ChainRulesCore
 else
     using ..DataInterpolations: _interpolate, derivative, AbstractInterpolation,
                                 LagrangeInterpolation, AkimaInterpolation,
                                 BSplineInterpolation, BSplineApprox, LinearInterpolation,
-                                linear_interpolation_parameters
+                                linear_interpolation_parameters, get_idx
     using ..ChainRulesCore
 end
+
+## Linear interpolation
 
 function ChainRulesCore.rrule(::typeof(linear_interpolation_parameters),
         u::AbstractVector, t::AbstractVector, idx::Integer)
     slope = linear_interpolation_parameters(u, t, idx)
-    Δt = t[idx + 1] - t[idx]
-    Δu = u[idx + 1] - t[idx]
     # TODO: use sparse arrays
     du = zero(u)
-    dt = zero(t)
 
     function linear_interpolation_parameters_pullback(Δ)
         df = NoTangent()
         du .= zero(eltype(u))
+        Δt = t[idx + 1] - t[idx]
         du[idx] = -Δ / Δt
         du[idx + 1] = Δ / Δt
-        dt .= zero(eltype(t))
-        dt[idx] = Δ * Δu / Δt^2
-        dt[idx + 1] = -Δ * Δu / Δt^2
+        dt = NoTangent()
         didx = NoTangent()
         return (df, du, dt, didx)
     end
@@ -45,30 +43,31 @@ function _tangent_u!(Δu::AbstractVector, A::LinearInterpolation, Δ)
     Δu
 end
 
-function _tangent_t!(Δt::AbstractVector, A::LinearInterpolation, Δ)
-    Δt .= zero(eltype(Δt))
-    Δt
-end
-
-function _tangent_p!(Δslope, A::LinearInterpolation, t, Δ)::Nothing
+function _tangent_p!(parameter_tangents::NamedTuple, A::LinearInterpolation, t, Δ)::Nothing
+    (; slope) = parameter_tangents
     idx = A.idx_prev[]
-    Δslope .= zero(eltype(A.p.slope))
-    Δslope[idx] = t * Δ
+    slope[idx] = (t - A.t[idx]) * Δ
     return nothing
 end
+
+function allocate_parameter_tangents(A::LinearInterpolation)
+    return (; slope = zero(A.p.slope))
+end
+
+## generic
 
 function ChainRulesCore.rrule(::typeof(_interpolate), A::AType, t) where {AType}
     u = _interpolate(A, t)
     # TODO: use sparse arrays
     Δu = zero(A.u)
-    Δt = zero(A.t)
-    Δslope = zero(A.p.slope)
+    parameter_tangents = allocate_parameter_tangents(A)
 
     function _interpolate_pullback(Δ)
+        A.idx_prev[] = get_idx(A.t, t, A.idx_prev[])
         df = NoTangent()
-        _tangent_p!(Δslope, A, t, Δ)
-        dA = Tangent{AType}(; u = _tangent_u!(Δu, A, Δ), t = _tangent_t!(Δt, A, Δ),
-            p = Tangent{typeof(A.p)}(; slope = Δslope))
+        _tangent_p!(parameter_tangents, A, t, Δ)
+        dA = Tangent{AType}(; u = _tangent_u!(Δu, A, Δ), t = NoTangent(),
[…]
