### 2024-07-09 406c76e
QuadraticSpline gradient w.r.t. u
