+++ src/parameter_caches.jl
@@ -4,11 +4,11 @@ end
 
 function LinearParameterCache(u, t)::LinearParameterCache
     idxs = 1:(length(t) - 1)
-    slope = [linear_interpolation_parameters(u, t, idx) for idx in idxs]
+    slope = [interpolation_parameters(Val(:LinearInterpolation), u, t, idx) for idx in idxs]
     return LinearParameterCache(slope)
 end
 
-function linear_interpolation_parameters(
+function interpolation_parameters(::Val{:LinearInterpolation},
         u::AbstractArray, t::AbstractVector, idx::Integer)
     Δu = u isa AbstractMatrix ? u[:, idx + 1] - u[:, idx] : u[idx + 1] - u[idx]
     Δt = t[idx + 1] - t[idx]
@@ -24,13 +24,13 @@ struct QuadraticParameterCache{pType}
 end
 
 function QuadraticParameterCache(u, t)
-    parameters = quadratic_interpolation_parameters.(
+    parameters = interpolation_parameters.(Val(:QuadraticInterpolation),
         Ref(u), Ref(t), 1:(length(t) - 2))
     l₀, l₁, l₂ = collect.(eachrow(hcat(collect.(parameters)...)))
     return QuadraticParameterCache(l₀, l₁, l₂)
 end
 
-function quadratic_interpolation_parameters(u, t, idx)
+function interpolation_parameters(::Val{:QuadraticInterpolation}, u, t, idx)
     if u isa AbstractMatrix
         u₀ = u[:, idx]
         u₁ = u[:, idx + 1]
@@ -57,11 +57,11 @@ struct QuadraticSplineParameterCache{pType}
 end
 
 function QuadraticSplineParameterCache(z, t)
-    σ = quadratic_spline_parameters.(Ref(z), Ref(t), 1:(length(t) - 1))
+    σ = interpolation_parameters.(Val(:QuadraticSpline), Ref(z), Ref(t), 1:(length(t) - 1))
     return QuadraticSplineParameterCache(σ)
 end
 
-function quadratic_spline_parameters(z, t, idx)
+function interpolation_parameters(::Val{:QuadraticSpline}, z, t, idx)
     σ = 1 // 2 * (z[idx + 1] - z[idx]) / (t[idx + 1] - t[idx])
     return σ
 end
@@ -72,13 +72,13 @@ struct CubicSplineParameterCache{pType}
 end
 
 function CubicSplineParameterCache(u, h, z)
-    parameters = cubic_spline_parameters.(
+    parameters = interpolation_parameters.(Val(:CubicSpline),
         Ref(u), Ref(h), Ref(z), 1:(size(u)[end] - 1))
     c₁, c₂ = collect.(eachrow(hcat(collect.(parameters)...)))
     return CubicSplineParameterCache(c₁, c₂)
 end
 
-function cubic_spline_parameters(u, h, z, idx)
+function interpolation_parameters(::Val{:CubicSpline}, u, h, z, idx)
     c₁ = (u[idx + 1] / h[idx + 1] - z[idx + 1] * h[idx + 1] / 6)
     c₂ = (u[idx] / h[idx + 1] - z[idx] * h[idx + 1] / 6)
     return c₁, c₂
@@ -90,13 +90,13 @@ struct CubicHermiteParameterCache{pType}
 end
 
 function CubicHermiteParameterCache(du, u, t)
-    parameters = cubic_hermite_spline_parameters.(
+    parameters = interpolation_parameters.(Val(:CubicHermiteSpline),
         Ref(du), Ref(u), Ref(t), 1:(length(t) - 1))
[…]
