+++ src/online.jl
@@ -3,7 +3,7 @@ import Base: append!, push!
 function push!(A::LinearInterpolation{U, T}, u::eltype(U), t::eltype(T)) where {U, T}
     push!(A.u.parent, u)
     push!(A.t.parent, t)
-    slope = linear_interpolation_parameters(A.u, A.t, length(A.t) - 1)
+    slope = interpolation_parameters(Val(:LinearInterpolation), A.u, A.t, length(A.t) - 1)
     push!(A.p.slope, slope)
     A
 end
@@ -11,7 +11,8 @@ end
 function push!(A::QuadraticInterpolation{U, T}, u::eltype(U), t::eltype(T)) where {U, T}
     push!(A.u.parent, u)
     push!(A.t.parent, t)
-    l₀, l₁, l₂ = quadratic_interpolation_parameters(A.u, A.t, length(A.t) - 2)
+    l₀, l₁, l₂ = interpolation_parameters(
+        Val(:QuadraticInterpolation), A.u, A.t, length(A.t) - 2)
     push!(A.p.l₀, l₀)
     push!(A.p.l₁, l₁)
     push!(A.p.l₂, l₂)
@@ -31,7 +32,7 @@ function append!(
     u, t = munge_data(u, t, true)
     append!(A.u.parent, u)
     append!(A.t.parent, t)
-    slope = linear_interpolation_parameters.(
+    slope = interpolation_parameters.(Val(:LinearInterpolation),
         Ref(A.u), Ref(A.t), length_old:(length(A.t) - 1))
     append!(A.p.slope, slope)
     A
@@ -53,7 +54,7 @@ function append!(
     u, t = munge_data(u, t, true)
     append!(A.u.parent, u)
     append!(A.t.parent, t)
-    parameters = quadratic_interpolation_parameters.(
+    parameters = interpolation_parameters.(Val(:QuadraticInterpolation),
         Ref(A.u), Ref(A.t), (length_old - 1):(length(A.t) - 2))
     l₀, l₁, l₂ = collect.(eachrow(hcat(collect.(parameters)...)))
     append!(A.p.l₀, l₀)
