+++ ext/DataInterpolationsChainRulesCoreExt.jl
@@ -1,75 +1,126 @@
 module DataInterpolationsChainRulesCoreExt
 
 if isdefined(Base, :get_extension)
-    using DataInterpolations: _interpolate, derivative, AbstractInterpolation,
+    using DataInterpolations: _interpolate, derivative, AbstractInterpolation, get_idx,
+                              interpolation_parameters,
                               LagrangeInterpolation, AkimaInterpolation,
                               BSplineInterpolation, BSplineApprox, LinearInterpolation,
-                              linear_interpolation_parameters, get_idx
+                              QuadraticSpline
     using ChainRulesCore
 else
-    using ..DataInterpolations: _interpolate, derivative, AbstractInterpolation,
+    using ..DataInterpolations: _interpolate, derivative, AbstractInterpolation, get_idx,
+                                interpolation_parameters,
                                 LagrangeInterpolation, AkimaInterpolation,
                                 BSplineInterpolation, BSplineApprox, LinearInterpolation,
-                                linear_interpolation_parameters, get_idx
+                                QuadraticSpline
     using ..ChainRulesCore
 end
 
 ## Linear interpolation
 
-function ChainRulesCore.rrule(::typeof(linear_interpolation_parameters),
+function ChainRulesCore.rrule(::typeof(interpolation_parameters),
+        ::Val{:LinearInterpolation},
         u::AbstractVector, t::AbstractVector, idx::Integer)
-    slope = linear_interpolation_parameters(u, t, idx)
+    slope = interpolation_parameters(Val(:LinearInterpolation), u, t, idx)
     # TODO: use sparse arrays
     du = zero(u)
+    Δt = t[idx + 1] - t[idx]
 
-    function linear_interpolation_parameters_pullback(Δ)
+    function interpolation_parameters_pullback(Δ)
         df = NoTangent()
-        du .= zero(eltype(u))
-        Δt = t[idx + 1] - t[idx]
+        dmethod = NoTangent()
         du[idx] = -Δ / Δt
         du[idx + 1] = Δ / Δt
         dt = NoTangent()
         didx = NoTangent()
-        return (df, du, dt, didx)
+        return (df, dmethod, du, dt, didx)
     end
 
-    return slope, linear_interpolation_parameters_pullback
+    return slope, interpolation_parameters_pullback
 end
 
-function _tangent_u!(Δu::AbstractVector, A::LinearInterpolation, Δ)
-    idx = A.idx_prev[]
-    Δu[idx] = Δ
-    Δu
-end
-
-function _tangent_p!(parameter_tangents::NamedTuple, A::LinearInterpolation, t, Δ)::Nothing
-    (; slope) = parameter_tangents
-    idx = A.idx_prev[]
-    slope[idx] = (t - A.t[idx]) * Δ
-    return nothing
+function allocate_direct_field_tangents(A::LinearInterpolation)
+    (; u = zero(A.u))
 end
 
 function allocate_parameter_tangents(A::LinearInterpolation)
     return (; slope = zero(A.p.slope))
 end
 
+function _tangent_direct_fields!(
+        direct_field_tangents::NamedTuple, A::LinearInterpolation, Δt, Δ)
+    (; u) = direct_field_tangents
+    idx = A.idx_prev[]
+    u[idx] = Δ
+end
+
+function _tangent_p!(parameter_tangents::NamedTuple, A::LinearInterpolation, Δt, Δ)
[…]
