### 2024-07-09 aa23c78
Use sparse arrays
