+++ src/interpolation_utils.jl
@@ -121,11 +121,7 @@ function get_idx(tvec, t, iguess; lb = 1, ub_shift = -1, idx_shift = 0, side = :
     end
 end
 
-function cumulative_integral(A)
-    if !(A.u isa AbstractVector{<:Number})
-        return nothing
-    end
-
+function cumulative_integral(A, ::AbstractVector{<:Number})
     integral_prototype = _integral(A, 1, A.t[2])
 
     integral_values = [zero(integral_prototype),
@@ -133,3 +129,5 @@ function cumulative_integral(A)
         for idx in 1:(length(A.t) - 1))...]
     return cumsum(integral_values)
 end
+
+cumulative_integral(A) = nothing
