### 2024-07-12 cae43d3
Throw extrapolation error also when interpolation is called with an `AbstractVector`
