+++ test/online_tests.jl
@@ -6,7 +6,8 @@ u1 = [0.0, 1.0, 0.0]
 t2 = [4.0, 5.0, 6.0]
 u2 = [1.0, 2.0, 1.0]
 
-ts = 1.0:0.5:6.0
+ts_append = 1.0:0.5:6.0
+ts_push = 1.0:0.5:4.0
 
 for method in [LinearInterpolation, QuadraticInterpolation, ConstantInterpolation]
     func1 = method(u1, t1)
@@ -17,7 +18,7 @@ for method in [LinearInterpolation, QuadraticInterpolation, ConstantInterpolatio
     for name in propertynames(func1.p)
         @test getfield(func1.p, name) == getfield(func2.p, name)
     end
-    @test func1(ts) == func2(ts)
+    @test func1(ts_append) == func2(ts_append)
 
     func1 = method(u1, t1)
     push!(func1, 1.0, 4.0)
@@ -27,5 +28,5 @@ for method in [LinearInterpolation, QuadraticInterpolation, ConstantInterpolatio
     for name in propertynames(func1.p)
         @test getfield(func1.p, name) == getfield(func2.p, name)
     end
-    @test func1(ts) == func2(ts)
+    @test func1(ts_push) == func2(ts_push)
 end
