+++ src/interpolation_caches.jl
@@ -1,5 +1,5 @@
 """
-    LinearInterpolation(u, t; extrapolate = false)
+    LinearInterpolation(u, t; extrapolate = false, safetycopy = true)
 
 It is the method of interpolating between the data points using a linear polynomial. For any point, two data points one each side are chosen and connected with a line.
 Extrapolation extends the last linear polynomial on each side.
@@ -37,7 +37,7 @@ function LinearInterpolation(u, t; extrapolate = false, safetycopy = true)
 end
 
 """
-    QuadraticInterpolation(u, t, mode = :Forward; extrapolate = false)
+    QuadraticInterpolation(u, t, mode = :Forward; extrapolate = false, safetycopy = true)
 
 It is the method of interpolating between the data points using quadratic polynomials. For any point, three data points nearby are taken to fit a quadratic polynomial.
 Extrapolation extends the last quadratic polynomial on each side.
@@ -83,7 +83,7 @@ function QuadraticInterpolation(u, t; extrapolate = false, safetycopy = true)
 end
 
 """
-    LagrangeInterpolation(u, t, n = length(t) - 1; extrapolate = false)
+    LagrangeInterpolation(u, t, n = length(t) - 1; extrapolate = false, safetycopy = true)
 
 It is the method of interpolation using Lagrange polynomials of (k-1)th order passing through all the data points where k is the number of data points.
 
@@ -134,7 +134,7 @@ function LagrangeInterpolation(
 end
 
 """
-    AkimaInterpolation(u, t; extrapolate = false)
+    AkimaInterpolation(u, t; extrapolate = false, safetycopy = true)
 
 It is a spline interpolation built from cubic polynomials. It forms a continuously differentiable function. For more details, refer: https://en.wikipedia.org/wiki/Akima_spline.
 Extrapolation extends the last cubic polynomial on each side.
@@ -203,7 +203,7 @@ function AkimaInterpolation(u, t; extrapolate = false, safetycopy = true)
 end
 
 """
-    ConstantInterpolation(u, t; dir = :left, extrapolate = false)
+    ConstantInterpolation(u, t; dir = :left, extrapolate = falseç safetycopy = true)
 
 It is the method of interpolating using a constant polynomial. For any point, two adjacent data points are found on either side (left and right). The value at that point depends on `dir`.
 If it is `:left`, then the value at the left point is chosen and if it is `:right`, the value at the right point is chosen.
@@ -243,7 +243,7 @@ function ConstantInterpolation(u, t; dir = :left, extrapolate = false, safetycop
 end
 
 """
-    QuadraticSpline(u, t; extrapolate = false)
+    QuadraticSpline(u, t; extrapolate = false, safetycopy = true)
 
 It is a spline interpolation using piecewise quadratic polynomials between each pair of data points. Its first derivative is also continuous.
 Extrapolation extends the last quadratic polynomial on each side.
@@ -329,7 +329,7 @@ function QuadraticSpline(
 end
 
 """
-    CubicSpline(u, t; extrapolate = false)
+    CubicSpline(u, t; extrapolate = false, safetycopy = true)
 
 It is a spline interpolation using piecewise cubic polynomials between each pair of data points. Its first and second derivative is also continuous.
 Second derivative on both ends are zero, which are also called "natural" boundary conditions. Extrapolation extends the last cubic polynomial on each side.
@@ -418,7 +418,7 @@ function CubicSpline(
 end
 
 """
-    BSplineInterpolation(u, t, d, pVecType, knotVecType; extrapolate = false)
[…]
