### 2024-07-13 9b0bbee
Add accelerated quasilinear index lookup for LinearInterpolation
