### 2024-07-13 12ebad8
Add user-facing mechanism to disable/enable linear-like guessing
