+++ ext/DataInterpolationsChainRulesCoreExt.jl
@@ -2,42 +2,53 @@ module DataInterpolationsChainRulesCoreExt
 
 if isdefined(Base, :get_extension)
     using DataInterpolations: _interpolate, derivative, AbstractInterpolation, get_idx,
-                              interpolation_parameters,
+                              interpolation_parameters, LinearParameterCache,
+                              QuadraticSplineParameterCache,
                               LagrangeInterpolation, AkimaInterpolation,
                               BSplineInterpolation, BSplineApprox, LinearInterpolation,
                               QuadraticSpline
     using ChainRulesCore
+    using LinearAlgebra
+    using SparseArrays
 else
     using ..DataInterpolations: _interpolate, derivative, AbstractInterpolation, get_idx,
-                                interpolation_parameters,
+                                interpolation_parameters, LinearParameterCache,
+                                QuadraticSplineParameterCache,
                                 LagrangeInterpolation, AkimaInterpolation,
                                 BSplineInterpolation, BSplineApprox, LinearInterpolation,
                                 QuadraticSpline
     using ..ChainRulesCore
+    using ..LinearAlgebra
+    using ..SparseArrays
 end
-
-using SparseArrays
 
 ## Linear interpolation
 
-function ChainRulesCore.rrule(::typeof(interpolation_parameters),
-        ::Val{:LinearInterpolation},
-        u::AbstractVector, t::AbstractVector, idx::Integer)
-    slope = interpolation_parameters(Val(:LinearInterpolation), u, t, idx)
-    du = SparseVector(length(u), [idx, idx + 1], zeros(2))
-    Δt = t[idx + 1] - t[idx]
+function ChainRulesCore.rrule(
+        ::Type{LinearParameterCache}, u::AbstractArray, t::AbstractVector)
+    p = LinearParameterCache(u, t)
+    du = zeros(eltype(p.slope), length(u))
 
-    function interpolation_parameters_pullback(Δ)
+    function LinearParameterCache_pullback(Δp)
         df = NoTangent()
-        dmethod = NoTangent()
-        du[idx] = -Δ / Δt
-        du[idx + 1] = Δ / Δt
+        du[2:end] += Δp.slope
+        du[1:(end - 1)] -= Δp.slope
         dt = NoTangent()
-        didx = NoTangent()
-        return (df, dmethod, du, dt, didx)
+        return (df, du, dt)
     end
 
-    return slope, interpolation_parameters_pullback
+    p, LinearParameterCache_pullback
+end
+
+function ChainRulesCore.rrule(
+        ::Type{LinearInterpolation}, u, t, I, p, extrapolate, safetycopy)
+    A = LinearInterpolation(u, t, I, p, extrapolate, safetycopy)
+
+    function LinearInterpolation_pullback(ΔA)
+        return ΔA.u, NoTangent(), NoTangent(), ΔA.p, NoTangent(), NoTangent(), NoTangent()
+    end
+
+    A, LinearInterpolation_pullback
 end
 
 function allocate_direct_field_tangents(A::LinearInterpolation)
@@ -67,50 +78,68 @@ end
 
 ## Quadratic Spline 
 
-function ChainRulesCore.rrule(::typeof(interpolation_parameters),
-        ::Val{:QuadraticSpline},
[…]
