+++ src/interpolation_utils.jl
@@ -107,7 +107,7 @@ function munge_data(U::StridedMatrix, t::AbstractVector, safetycopy::Bool)
 end
 
 # Don't nest ReadOnlyArrays
-readonly_wrap(a::AbstractArray) = a
+readonly_wrap(a::AbstractArray) = ReadOnlyArray(a)
 readonly_wrap(a::ReadOnlyArray) = a
 
 function get_idx(tvec, t, iguess; lb = 1, ub_shift = -1, idx_shift = 0, side = :last)
@@ -130,4 +130,4 @@ function cumulative_integral(A, ::AbstractVector{<:Number})
     return cumsum(integral_values)
 end
 
-cumulative_integral(A) = nothing
+cumulative_integral(A, ::AbstractArray) = nothing
