+++ Project.toml
@@ -39,6 +39,7 @@ RecipesBase = "1.3"
 Reexport = "1"
 RegularizationTools = "0.6"
 SafeTestsets = "0.1"
+SparseArrays = "1.10"
 StableRNGs = "1"
 Symbolics = "5.29"
 Test = "1"

[…]
