### 2024-07-13 30c99c7
Add ReadOnlyArray rrule
