+++ ext/DataInterpolationsChainRulesCoreExt.jl
@@ -2,7 +2,7 @@ module DataInterpolationsChainRulesCoreExt
 
 if isdefined(Base, :get_extension)
     using DataInterpolations: _interpolate, derivative, AbstractInterpolation, get_idx,
-                              interpolation_parameters, LinearParameterCache,
+                              cumulative_integral, LinearParameterCache,
                               QuadraticSplineParameterCache,
                               LagrangeInterpolation, AkimaInterpolation,
                               BSplineInterpolation, BSplineApprox, LinearInterpolation,
@@ -10,9 +10,10 @@ if isdefined(Base, :get_extension)
     using ChainRulesCore
     using LinearAlgebra
     using SparseArrays
+    using ReadOnlyArrays
 else
     using ..DataInterpolations: _interpolate, derivative, AbstractInterpolation, get_idx,
-                                interpolation_parameters, LinearParameterCache,
+                                cumulative_integral, LinearParameterCache,
                                 QuadraticSplineParameterCache,
                                 LagrangeInterpolation, AkimaInterpolation,
                                 BSplineInterpolation, BSplineApprox, LinearInterpolation,
@@ -20,6 +21,7 @@ else
     using ..ChainRulesCore
     using ..LinearAlgebra
     using ..SparseArrays
+    using ..ReadOnlyArrays
 end
 
 ## Linear interpolation
@@ -45,7 +47,14 @@ function ChainRulesCore.rrule(
     A = LinearInterpolation(u, t, I, p, extrapolate, safetycopy)
 
     function LinearInterpolation_pullback(ΔA)
-        return ΔA.u, NoTangent(), NoTangent(), ΔA.p, NoTangent(), NoTangent(), NoTangent()
+        df = NoTangent()
+        du = ΔA.u
+        dt = NoTangent()
+        dI = NoTangent()
+        dp = ΔA.p
+        dextrapolate = NoTangent()
+        dsafetycopy = NoTangent()
+        return df, du, dt, dI, dp, dextrapolate, dsafetycopy
     end
 
     A, LinearInterpolation_pullback
@@ -82,8 +91,6 @@ function ChainRulesCore.rrule(::Type{QuadraticSplineParameterCache}, u, t)
     p = QuadraticSplineParameterCache(u, t)
     n = length(u)
 
-    ∂z_∂d = inv(p.tA)
-
     Δt = diff(t)
     diagonal_main = [zero(eltype(Δt)), 2 ./ Δt...]
     diagonal_down = -diagonal_main[2:end]
@@ -98,7 +105,9 @@ function ChainRulesCore.rrule(::Type{QuadraticSplineParameterCache}, u, t)
 
     function QuadraticSplineParameterCache_pullback(Δp)
         df = NoTangent()
-        du = (Δp.z + ∂σ_∂z * Δp.σ)' * ∂z_∂d * ∂d_∂u
+        temp1 = Δp.z + ∂σ_∂z * Δp.σ
+        temp2 = p.tA' \ temp1
+        du = ∂d_∂u' * temp2
         dt = NoTangent()
         return (df, du, dt)
     end
@@ -109,11 +118,18 @@ end
 function ChainRulesCore.rrule(::Type{QuadraticSpline}, u, t, I, p, extrapolate, safetycopy)
     A = QuadraticSpline(u, t, I, p, extrapolate, safetycopy)
 
-    function LinearInterpolation_pullback(ΔA)
-        return ΔA.u, NoTangent(), NoTangent(), ΔA.p, NoTangent(), NoTangent(), NoTangent()
+    function QuadraticSpline_pullback(ΔA)
+        df = NoTangent()
[…]
