+++ src/DataInterpolations.jl
@@ -8,7 +8,6 @@ using LinearAlgebra, RecipesBase
 using PrettyTables
 using ForwardDiff
 using ReadOnlyArrays
-using SparseArrays # Only used in DataInterpolationsChainRulesCoreExt.jl, but otherwise Aqua complains
 import FindFirstFunctions: searchsortedfirstcorrelated, searchsortedlastcorrelated,
                            bracketstrictlymontonic
 
