### 2024-07-14 985e454
Tighten quasi-regularity criterion
