+++ src/interpolation_caches.jl
@@ -24,10 +24,11 @@ struct LinearInterpolation{uType, tType, IType, pType, T} <: AbstractInterpolati
     safetycopy::Bool
     seems_linear::Bool
     function LinearInterpolation(
-            u, t, I, p, extrapolate, safetycopy; guess_linear_t = nothing)
-        guess_linear_t = isnothing(guess_linear_t) ? looks_quasilinear(t) : guess_linear_t
+            u, t, I, p, extrapolate, safetycopy; quasiregular = 1e-2)
+        quasiregular = quasiregular isa Bool ? quasiregular :
+                       looks_quasiregular(t; threshold = quasiregular)
         new{typeof(u), typeof(t), typeof(I), typeof(p.slope), eltype(u)}(
-            u, t, I, p, extrapolate, Ref(1), safetycopy, looks_quasilinear(t))
+            u, t, I, p, extrapolate, Ref(1), safetycopy, quasiregular)
     end
 end
 
