### 2024-07-15 a9c3624
docs: small cleanup of pchip docstrings
