### 2024-07-16 #6917 @carstenbauer
I've made my pass over my checklist, the package source code/documentation, and the short paper. Here are my comments and questions. All of them should be straightforward to integrate and answer, respectively.

* The comparison to other interpolation packages in Julia (most notably [Interpolations.jl](https://github.com/JuliaMath/Interpolations.jl)) could be improved. First of all, these other options should be mentioned explicitly with links (at least [Interpolations.jl](https://github.com/JuliaMath/Interpolations.jl)). Secondly, what are the differences? The paper mentions that other packages require uniformly spaced data but at least according to the README.md of [Interpolations.jl](https://github.com/JuliaMath/Interpolations.jl) they also support irregular grids.
* The target audience could be clarified more clearly (perhaps even addressed explicitly) in the statement of need section. A single sentence is sufficient.
* The final sentences of the statement of need sections of the paper - especially the part about the symbiosis with Symbolics.jl and ModelingToolkit.jl - would perhaps be better suited in the summary section.
* There is no statement of need section in the documentation. I'm fine with the documentation as is but the JOSS checklist asks for it, which is why I bring it up here.

**Bonus:**
* It would be neat if the example would also highlight interoperability with Symbolics.jl, especially because this feature is mentioned in the text and could be a distinguishing factor of the package in comparison to other packages (in Julia and other languages).
* The quality of writing is okay, but could probably be improved a little with little effort.

Provided that these points are taken into account by the authors, I am happy to recommend the paper for publication.
