### 2024-07-17 4aff70c
docs: changes in joss paper according to reviews
