### 2024-07-17 601491d
docs: include all suggestions for joss paper
