### 2024-07-17 558c683
build(docs): add Symbolics, MTK and reqired deps
