### 2024-07-22 69dbed7
fix: type inference for interface
The interface for, e.g. LinearInterpolation, is not type stable because
`cummulative_integral` returns a union. This is because it does a
runtime `isempty` check on the methods table. Replacing this with a
`hasmethod` check allows the boolean to be determined during
compilation, fixing the types.
