+++ test/interface.jl
@@ -1,6 +1,7 @@
 using DataInterpolations
 u = 2.0collect(1:10)
 t = 1.0collect(1:10)
+@inferred LinearInterpolation(u, t)
 A = LinearInterpolation(u, t)
 
 for i in 1:10
