+++ .github/workflows/Downgrade.yml
@@ -28,7 +28,7 @@ jobs:
           - windows-latest
     steps:
       - uses: actions/checkout@v4
-      - uses: julia-actions/setup-julia@v2.2.0
+      - uses: julia-actions/setup-julia@v2.3.0
         with:
           version: ${{ matrix.version }}
       - uses: julia-actions/julia-downgrade-compat@v1
