### 2024-07-23 c80d7bd
refactor: collect parameters such that they are type stable
