### 2024-07-23 826f0a7
refactor: use stack instead of reduce hcat
