+++ src/parameter_caches.jl
@@ -7,8 +7,19 @@ function LinearParameterCache(u, t)
     return LinearParameterCache(slope)
 end
 
-function linear_interpolation_parameters(u, t, idx)
-    Δu = u isa AbstractMatrix ? u[:, idx + 1] - u[:, idx] : u[idx + 1] - u[idx]
+"""
+Prevent e.g. Inf - Inf = NaN
+"""
+function safe_diff(b, a::T) where T
+    b == a ? zero(T) : b - a 
+end
+
+function linear_interpolation_parameters(u::AbstractArray{T}, t, idx) where T
+    Δu = if u isa AbstractMatrix
+        [safe_diff(u[j, idx + 1], u[j, idx]) for j in 1:size(u)[1]]
+    else
+        safe_diff(u[idx + 1], u[idx]) 
+    end
     Δt = t[idx + 1] - t[idx]
     slope = Δu / Δt
     slope = iszero(Δt) ? zero(slope) : slope
