### 2024-07-23 485dc32
Remove docstring
