+++ dev/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"inverting_integrals/#Inverting-integrals","page":"Inverting Integrals","title":"Inverting integrals","text":"","category":"section"},{"location":"inverting_integrals/","page":"Inverting Integrals","title":"Inverting Integrals","text":"Solving implicit integral problems of the following form is supported:","category":"page"},{"location":"inverting_integrals/","page":"Inverting Integrals","title":"Inverting Integrals","text":"beginequation\n    textfind t such that  int_t_1^t f(tau)textdtau = V ge 0\nendequation","category":"page"},{"location":"inverting_integrals/","page":"Inverting Integrals","title":"Inverting Integrals","text":"where t_1 is given by first(A.t). This is supported for interpolations f that are strictly positive and of one of these types:","category":"page"},{"location":"inverting_integrals/","page":"Inverting Integrals","title":"Inverting Integrals","text":"ConstantInterpolation\nLinearInterpolation","category":"page"},{"location":"inverting_integrals/","page":"Inverting Integrals","title":"Inverting Integrals","text":"This is achieved by creating an 'integral inverse' interpolation object which can efficiently compute t for a given value of V, see the example below.","category":"page"},{"location":"inverting_integrals/","page":"Inverting Integrals","title":"Inverting Integrals","text":"using Random #hide\nRandom.seed!(1234) # hide\nusing DataInterpolations\nusing Plots\n\n# Create LinearInterpolation object from the\nu = sqrt.(1:25) + (2.0 * rand(25) .- 1.0) / 3\nt = cumsum(rand(25))\nA = LinearInterpolation(u, t)\n\n# Create LinearInterpolationIntInv object\n# from the LinearInterpolation object\nA_intinv = DataInterpolations.invert_integral(A)\n\n# Get the t values up to and including the\n# solution to the integral problem\nV = 25.0\nt_ = A_intinv(V)\nts = t[t .<= t_]\npush!(ts, t_)\n\n# Plot results\nplot(A; label = \"Linear Interpolation\")\nplot!(ts, A.(ts), fillrange = 0.0, fillalpha = 0.75,\n    fc = :blues, lw = 0, label = \"Area of $V\")","category":"page"},{"location":"inverting_integrals/#Docstrings","page":"Inverting Integrals","title":"Docstrings","text":"","category":"section"},{"location":"inverting_integrals/","page":"Inverting Integrals","title":"Inverting Integrals","text":"DataInterpolations.invert_integral\nConstantInterpolationIntInv\nLinearInterpolationIntInv","category":"page"},{"location":"inverting_integrals/#DataInterpolations.invert_integral","page":"Inverting Integrals","title":"DataInterpolations.invert_integral","text":"invert_integral(A::AbstractInterpolation)::AbstractIntegralInverseInterpolation\n\nCreates the inverted integral interpolation object from the given interpolation. Conditions:\n\nThe range of A must be strictly positive\nA.u must be a number type (on which an ordering is defined)\nThis is currently only supported for ConstantInterpolation and LinearInterpolation\n\nArguments\n\nA: interpolation object satisfying the above requirements\n\n\n\n\n\n","category":"function"},{"location":"inverting_integrals/#DataInterpolations.ConstantInterpolationIntInv","page":"Inverting Integrals","title":"DataInterpolations.ConstantInterpolationIntInv","text":"ConstantInterpolationIntInv(u, t, A)\n\nIt is the interpolation of the inverse of the integral of a ConstantInterpolation. Can be easily constructed with invert_integral(A::ConstantInterpolation{<:AbstractVector{<:Number}})\n\nArguments\n\
[…]
