+++ versions.js
@@ -13,5 +13,5 @@ var DOC_VERSIONS = [
   "v4.1",
   "dev",
 ];
-var DOCUMENTER_NEWEST = "v5.3.0";
+var DOCUMENTER_NEWEST = "v5.3.1";
 var DOCUMENTER_STABLE = "stable";
