### 2024-07-26 #6917 @sathvikbhagavan
Thanks @carstenbauer for the review!

I have addressed 1, 2, 3 points in the paper. I have addressed 4 - added statement of need in the documentation. I have also added a tutorial for using DataInterpolations with Symbolics and referenced that in the paper.
Do take a look if any other further changes are needed.
