### 2024-07-27 c8b3224
Refactor parameter caching, add zygote tests
