+++ test/interpolation_tests.jl
@@ -9,7 +9,6 @@ function test_interpolation_type(T)
     @test hasfield(T, :t)
     @test hasfield(T, :extrapolate)
     @test hasfield(T, :idx_prev)
-    @test hasfield(T, :safetycopy)
     @test !isempty(methods(DataInterpolations._interpolate, (T, Any, Number)))
     @test !isempty(methods(DataInterpolations._integral, (T, Any, Number)))
     @test !isempty(methods(DataInterpolations._derivative, (T, Any, Number)))
