+++ src/interpolation_caches.jl
@@ -1,5 +1,5 @@
 """
-    LinearInterpolation(u, t; extrapolate = false)
+    LinearInterpolation(u, t; extrapolate = false, cache_parameters = false)
 
 It is the method of interpolating between the data points using a linear polynomial. For any point, two data points one each side are chosen and connected with a line.
 Extrapolation extends the last linear polynomial on each side.
@@ -12,7 +12,7 @@ Extrapolation extends the last linear polynomial on each side.
 ## Keyword Arguments
 
   - `extrapolate`: boolean value to allow extrapolation. Defaults to `false`.
-  - `safetycopy`: boolean value to make a copy of `u` and `t`. Defaults to `true`.
+  - `cache_parameters`: precompute parameters at initialization for faster interpolation computations. Note: if activated, `u` and `t` should not be modified. Defaults to `false`.
 """
 struct LinearInterpolation{uType, tType, IType, pType, T} <: AbstractInterpolation{T}
     u::uType
@@ -21,23 +21,33 @@ struct LinearInterpolation{uType, tType, IType, pType, T} <: AbstractInterpolati
     p::LinearParameterCache{pType}
     extrapolate::Bool
     idx_prev::Base.RefValue{Int}
-    safetycopy::Bool
-    function LinearInterpolation(u, t, I, p, extrapolate, safetycopy)
+    cache_parameters::Bool
+    function LinearInterpolation(u, t, I, p, extrapolate, cache_parameters)
         new{typeof(u), typeof(t), typeof(I), typeof(p.slope), eltype(u)}(
-            u, t, I, p, extrapolate, Ref(1), safetycopy)
+            u, t, I, p, extrapolate, Ref(1), cache_parameters)
     end
 end
 
-function LinearInterpolation(u, t; extrapolate = false, safetycopy = true)
-    u, t = munge_data(u, t, safetycopy)
-    p = LinearParameterCache(u, t)
-    A = LinearInterpolation(u, t, nothing, p, extrapolate, safetycopy)
-    I = cumulative_integral(A)
-    LinearInterpolation(u, t, I, p, extrapolate, safetycopy)
+function LinearInterpolation(u, t; extrapolate = false, cache_parameters = false)
+    u, t = munge_data(u, t)
+    p = if cache_parameters
+        LinearParameterCache(u, t)
+    else
+        LinearParameterCache(nothing)
+    end
+
+    A = LinearInterpolation(u, t, nothing, p, extrapolate, cache_parameters)
+
+    if cache_parameters
+        I = cumulative_integral(A)
+        A = LinearInterpolation(u, t, I, p, extrapolate, cache_parameters)
+    end
+
+    A
 end
 
 """
-    QuadraticInterpolation(u, t, mode = :Forward; extrapolate = false)
+    QuadraticInterpolation(u, t, mode = :Forward; cache_parameters = false)
 
 It is the method of interpolating between the data points using quadratic polynomials. For any point, three data points nearby are taken to fit a quadratic polynomial.
 Extrapolation extends the last quadratic polynomial on each side.
@@ -51,7 +61,7 @@ Extrapolation extends the last quadratic polynomial on each side.
 ## Keyword Arguments
 
   - `extrapolate`: boolean value to allow extrapolation. Defaults to `false`.
-  - `safetycopy`: boolean value to make a copy of `u` and `t`. Defaults to `true`.
+  - `cache_parameters`: precompute parameters at initialization for faster interpolation computations. Note: if activated, `u` and `t` should not be modified. Defaults to `false`.
 """
[…]
