+++ src/DataInterpolations.jl
@@ -7,7 +7,6 @@ abstract type AbstractInterpolation{T} end
 using LinearAlgebra, RecipesBase
 using PrettyTables
 using ForwardDiff
-using ReadOnlyArrays
 import FindFirstFunctions: searchsortedfirstcorrelated, searchsortedlastcorrelated,
                            bracketstrictlymontonic
 
@@ -88,12 +87,6 @@ function Base.showerror(io::IO, e::IntegralNotInvertibleError)
     print(io, INTEGRAL_NOT_INVERTIBLE_ERROR)
 end
 
-const MUST_COPY_ERROR = "A copy must be made of u, t to filter missing data"
-struct MustCopyError <: Exception end
-function Base.showerror(io::IO, e::MustCopyError)
-    print(io, MUST_COPY_ERROR)
-end
-
 export LinearInterpolation, QuadraticInterpolation, LagrangeInterpolation,
        AkimaInterpolation, ConstantInterpolation, QuadraticSpline, CubicSpline,
        BSplineInterpolation, BSplineApprox, CubicHermiteSpline,
@@ -126,12 +119,12 @@ struct RegularizationSmooth{uType, tType, T, T2, ITP <: AbstractInterpolation{T}
             Aitp,
             extrapolate)
         new{typeof(u), typeof(t), eltype(u), typeof(λ), typeof(Aitp)}(
-            readonly_wrap(u),
-            readonly_wrap(û),
-            readonly_wrap(t),
-            readonly_wrap(t̂),
-            readonly_wrap(oftype(u.parent, wls)),
-            readonly_wrap(oftype(u.parent, wr)),
+            u,
+            û,
+            t,
+            t̂,
+            wls,
+            wr,
             d,
             λ,
             alg,

[…]
