+++ test/interpolation_tests.jl
@@ -160,6 +160,13 @@ end
     @test A(5.5) == fill(11.0)
     @test A(11) == fill(22)
 
+    # Test constant -Inf interpolation
+    u = [-Inf, -Inf]
+    t = [0.0, 1.0]
+    A = LinearInterpolation(u, t)
+    @test A(0.0) == -Inf
+    @test A(0.5) == -Inf
+
     # Test extrapolation
     u = 2.0collect(1:10)
     t = 1.0collect(1:10)
@@ -169,6 +176,7 @@ end
     A = LinearInterpolation(u, t)
     @test_throws DataInterpolations.ExtrapolationError A(-1.0)
     @test_throws DataInterpolations.ExtrapolationError A(11.0)
+    @test_throws DataInterpolations.ExtrapolationError A([-1.0, 11.0])
 end
 
 @testset "Quadratic Interpolation" begin
@@ -667,6 +675,18 @@ end
     test_cached_index(A)
     push!(u, 1.0)
     @test_throws AssertionError CubicHermiteSpline(du, u, t)
+end
+
+@testset "PCHIPInterpolation" begin
+    u = [14.7, 11.51, 10.41, 14.95, 12.24, 11.22]
+    t = [0.0, 62.25, 109.66, 162.66, 205.8, 250.0]
+    A = PCHIPInterpolation(u, t)
+    @test A isa CubicHermiteSpline
+    ts = 0.0:0.1:250.0
+    us = A(ts)
+    @test all(minimum(u) .<= us)
+    @test all(maximum(u) .>= us)
+    @test all(A.du[3:4] .== 0.0)
 end
 
 @testset "Quintic Hermite Spline" begin
