+++ Project.toml
@@ -39,7 +39,7 @@ SafeTestsets = "0.1"
 StableRNGs = "1"
 Symbolics = "5.29"
 Test = "1"
-Zygote = "^0"
+Zygote = "0.6"
 julia = "1.10"
 
 [extras]
