### 2024-07-28 e0cce0b
Add examples of how to speed up gradients w.r.t. u, add more zygote tests
