+++ test/zygote_tests.jl
@@ -2,8 +2,8 @@ using DataInterpolations
 using ForwardDiff
 using Zygote
 
-function test_zygote(method, u, t; args = [], kwargs = [], name::String)
-    func = method(args..., u, t; kwargs..., extrapolate = true)
+function test_zygote(method, u, t; args = [], args_after = [], kwargs = [], name::String)
+    func = method(args..., u, t, args_after...; kwargs..., extrapolate = true)
     (; u, t) = func
     trange = collect(range(minimum(t) - 5.0, maximum(t) + 5.0, step = 0.1))
     trange_exclude = filter(x -> !in(x, t), trange)
@@ -11,28 +11,30 @@ function test_zygote(method, u, t; args = [], kwargs = [], name::String)
         for _t in trange_exclude
             adiff = DataInterpolations.derivative(func, _t)
             zdiff = only(Zygote.gradient(func, _t))
-            zdiff == nothing && (zdiff = 0.0)
+            isnothing(zdiff) && (zdiff = 0.0)
             @test adiff ≈ zdiff
         end
     end
-    @testset "$name, derivatives w.r.t. u" begin
-        function f(u)
-            A = method(args..., u, t; kwargs..., extrapolate = true)
-            out = zero(eltype(u))
-            for _t in trange
-                out += A(_t)
+    if method ∉ [LagrangeInterpolation, BSplineInterpolation, BSplineApprox]
+        @testset "$name, derivatives w.r.t. u" begin
+            function f(u)
+                A = method(args..., u, t, args_after...; kwargs..., extrapolate = true)
+                out = zero(eltype(u))
+                for _t in trange
+                    out += A(_t)
+                end
+                out
             end
-            out
+            zgrad = only(Zygote.gradient(f, u))
+            fgrad = ForwardDiff.gradient(f, u)
+            @test zgrad ≈ fgrad
         end
-        zgrad = only(Zygote.gradient(f, u))
-        fgrad = ForwardDiff.gradient(f, u)
-        @test zgrad ≈ fgrad
     end
 end
 
 @testset "LinearInterpolation" begin
-    u = vcat(collect(1:5), 2 * collect(6:10))
-    t = 1.0collect(1:10)
+    u = vcat(collect(1.0:5.0), 2 * collect(6.0:10.0))
+    t = collect(1.0:10.0)
     test_zygote(
         LinearInterpolation, u, t; name = "Linear Interpolation")
 end
@@ -63,4 +65,37 @@ end
     t = [0.0, 62.25, 109.66, 162.66, 205.8, 252.3]
     test_zygote(
         QuinticHermiteSpline, u, t, args = [ddu, du], name = "Quintic Hermite Spline")
+end
+
+@testset "Quadratic Spline" begin
+    u = [1.0, 4.0, 9.0, 16.0]
+    t = [1.0, 2.0, 3.0, 4.0]
+    test_zygote(QuadraticSpline, u, t, name = "Quadratic Spline")
+end
+
+@testset "Lagrange Interpolation" begin
+    u = [1.0, 4.0, 9.0]
+    t = [1.0, 2.0, 3.0]
+    test_zygote(LagrangeInterpolation, u, t, name = "Lagrange Interpolation")
+end
+
+@testset "Constant Interpolation" begin
+    u = [0.0, 2.0, 1.0, 3.0, 2.0, 6.0, 5.5, 5.5, 2.7, 5.1, 3.0]
+    t = collect(0.0:10.0)
+    test_zygote(ConstantInterpolation, u, t, name = "Constant Interpolation")
+end
+
+@testset "Cubic Spline" begin
[…]
