### 2024-07-29 bf37761
build: bump zygote compat to get downgrade tests working
