+++ Project.toml
@@ -40,7 +40,7 @@ SafeTestsets = "0.1"
 StableRNGs = "1"
 Symbolics = "5.29"
 Test = "1"
-Zygote = "0.6"
+Zygote = "0.6.70"
 julia = "1.10"
 
 [extras]
