### 2024-07-29 21516aa
build: bump ChainRulesCore compat
