### 2024-07-29 245fe51
test: add tests for symbolic interpolation compositions
