### 2024-07-29 be7366c
build(docs): bump DataInterpolations compat
