### 2024-07-29 2ab91ed
[WIP] Wording, toggling settings for linear lookup
