+++ src/interpolation_utils.jl
@@ -113,7 +113,17 @@ function looks_linear(t; threshold = 1e-2)
     norm_var < threshold^2
 end
 
-function get_idx(tvec, t, iguess; lb = 1, ub_shift = -1, idx_shift = 0, side = :last)
+function get_idx(A::AbstractInterpolation, t, iguess; lb = 1,
+        ub_shift = -1, idx_shift = 0, side = :last)
+    iguess = if hasfield(typeof(A), :use_linear_lookup) &&
+                A.use_linear_lookup
+        f = (t - first(A.t)) / (last(A.t) - first(A.t))
+        i_0, i_f = firstindex(A.t), lastindex(A.t)
+        round(typeof(firstindex(A.t)), f * (i_f - i_0) + i_0)
+    else
+        iguess
+    end
+    tvec = A.t
     ub = length(tvec) + ub_shift
     return if side == :last
         clamp(searchsortedlastcorrelated(tvec, t, iguess) + idx_shift, lb, ub)
