+++ src/interpolation_caches.jl
@@ -29,9 +29,9 @@ struct LinearInterpolation{uType, tType, IType, pType, T} <: AbstractInterpolati
     cache_parameters::Bool
     use_linear_lookup::Bool
     function LinearInterpolation(u, t, I, p, extrapolate, cache_parameters, assume_linear_t)
-        linear_flag = seems_linear(assume_linear_t, t)
+        linear_lookup = seems_linear(assume_linear_t, t)
         new{typeof(u), typeof(t), typeof(I), typeof(p.slope), eltype(u)}(
-            u, t, I, p, extrapolate, Ref(1), cache_parameters, linear_flag)
+            u, t, I, p, extrapolate, Ref(1), cache_parameters, linear_lookup)
     end
 end
 
@@ -39,7 +39,8 @@ function LinearInterpolation(
         u, t; extrapolate = false, cache_parameters = false, assume_linear_t = 1e-2)
     u, t = munge_data(u, t)
     p = LinearParameterCache(u, t, cache_parameters)
-    A = LinearInterpolation(u, t, nothing, p, extrapolate, cache_parameters)
+    A = LinearInterpolation(
+        u, t, nothing, p, extrapolate, cache_parameters, assume_linear_t)
     I = cumulative_integral(A, cache_parameters)
     LinearInterpolation(u, t, I, p, extrapolate, cache_parameters, assume_linear_t)
 end
@@ -60,6 +61,10 @@ Extrapolation extends the last quadratic polynomial on each side.
 
   - `extrapolate`: boolean value to allow extrapolation. Defaults to `false`.
   - `cache_parameters`: precompute parameters at initialization for faster interpolation computations. Note: if activated, `u` and `t` should not be modified. Defaults to `false`.
+  - `assume_linear_t`: boolean value to specify a faster index lookup behaviour for
+    evenly-distributed abscissae. Alternatively, a numerical threshold may be specified
+    for a test based on the normalized standard deviation of the difference with respect
+    to the straight line. Defaults to 1e-2
 """
 struct QuadraticInterpolation{uType, tType, IType, pType, T} <: AbstractInterpolation{T}
     u::uType
@@ -75,22 +80,25 @@ struct QuadraticInterpolation{uType, tType, IType, pType, T} <: AbstractInterpol
             u, t, I, p, mode, extrapolate, cache_parameters, assume_linear_t)
         mode ∈ (:Forward, :Backward) ||
             error("mode should be :Forward or :Backward for QuadraticInterpolation")
-        linear_flag = seems_linear(assume_linear_t, t)
+        linear_lookup = seems_linear(assume_linear_t, t)
         new{typeof(u), typeof(t), typeof(I), typeof(p.l₀), eltype(u)}(
-            u, t, I, p, mode, extrapolate, Ref(1), cache_parameters, linear_flag)
+            u, t, I, p, mode, extrapolate, Ref(1), cache_parameters, linear_lookup)
     end
 end
 
-function QuadraticInterpolation(u, t, mode; extrapolate = false, cache_parameters = false)
+function QuadraticInterpolation(
+        u, t, mode; extrapolate = false, cache_parameters = false, assume_linear_t = 1e-2)
     u, t = munge_data(u, t)
+    linear_lookup = seems_linear(assume_linear_t, t)
     p = QuadraticParameterCache(u, t, cache_parameters)
-    A = QuadraticInterpolation(u, t, nothing, p, mode, extrapolate, cache_parameters)
+    A = QuadraticInterpolation(
+        u, t, nothing, p, mode, extrapolate, cache_parameters, linear_lookup)
     I = cumulative_integral(A, cache_parameters)
[…]
