+++ src/interpolation_utils.jl
@@ -97,11 +97,11 @@ end
 seems_linear(assume_linear_t::Bool, _) = assume_linear_t
 seems_linear(assume_linear_t::Number, t) = looks_linear(t; threshold = assume_linear_t)
 
-# """
-# Determine if the abscissae are regularly distributed, taking the standard \
-# deviation of the difference between the array of abscissae with respect to \
-# the straight line linking its first and last elements
-# """
+"""
+Determine if the abscissae `t` are regularly distributed, taking the standard deviation of \
+the difference between the array of abscissae with respect to the straight line linking \
+its first and last elements, normalized by the range of `t`.
+"""
 function looks_linear(t; threshold = 1e-2)
     length(t) <= 2 && return true
     t_0, t_f = first(t), last(t)
@@ -115,8 +115,8 @@ end
 
 function get_idx(A::AbstractInterpolation, t, iguess; lb = 1,
         ub_shift = -1, idx_shift = 0, side = :last)
-    iguess = if hasfield(typeof(A), :use_linear_lookup) &&
-                A.use_linear_lookup
+    iguess = if hasfield(typeof(A), :linear_lookup) &&
+                A.linear_lookup
         f = (t - first(A.t)) / (last(A.t) - first(A.t))
         i_0, i_f = firstindex(A.t), lastindex(A.t)
         round(typeof(firstindex(A.t)), f * (i_f - i_0) + i_0)
