### 2024-07-30 4380c66
Update all docstrings to reference looks_linear
