### 2024-07-31 179f1db
Docstring changes for looks_linear
