+++ previews/PR321/assets/Project.toml
@@ -1,0 +1,23 @@
+[deps]
+DataInterpolations = "82cc6244-b520-54b8-b5a6-8a565e85f1d0"
+Documenter = "e30172f5-a6a5-5a46-863b-614d45cd2de4"
+ModelingToolkit = "961ee093-0014-501f-94e3-6117800e7a78"
+ModelingToolkitStandardLibrary = "16a59e39-deab-5bd0-87e4-056b12336739"
+Optim = "429524aa-4258-5aef-a3af-852621145aeb"
+OrdinaryDiffEq = "1dea7af3-3e70-54e6-95c3-0bf5283fa5ed"
+Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
+RegularizationTools = "29dad682-9a27-4bc3-9c72-016788665182"
+StableRNGs = "860ef19b-820b-49d6-a774-d7a799459cd3"
+Symbolics = "0c5d862f-8b57-4792-8d23-62f2024744c7"
+
+[compat]
+DataInterpolations = "6"
+Documenter = "1"
+ModelingToolkit = "9"
+ModelingToolkitStandardLibrary = "2"
+Optim = "1"
+OrdinaryDiffEq = "6"
+Plots = "1"
+RegularizationTools = "0.6"
+StableRNGs = "1"
+Symbolics = "5.29"
