+++ dev/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"symbolics/#Using-DataInterpolations.jl-with-Symbolics.jl-and-ModelingToolkit.jl","page":"Using with Symbolics/ModelingToolkit","title":"Using DataInterpolations.jl with Symbolics.jl and ModelingToolkit.jl","text":"","category":"section"},{"location":"symbolics/","page":"Using with Symbolics/ModelingToolkit","title":"Using with Symbolics/ModelingToolkit","text":"All interpolation methods can be integrated with Symbolics.jl and ModelingToolkit.jl seamlessly.","category":"page"},{"location":"symbolics/#Using-with-Symbolics.jl","page":"Using with Symbolics/ModelingToolkit","title":"Using with Symbolics.jl","text":"","category":"section"},{"location":"symbolics/#Expressions","page":"Using with Symbolics/ModelingToolkit","title":"Expressions","text":"","category":"section"},{"location":"symbolics/","page":"Using with Symbolics/ModelingToolkit","title":"Using with Symbolics/ModelingToolkit","text":"using DataInterpolations, Symbolics\nusing Test\n\nu = [0.0, 1.5, 0.0]\nt = [0.0, 0.5, 1.0]\nA = LinearInterpolation(u, t)\n\n@variables τ\n\n# Simple Expression\nex = cos(τ) * A(τ)\n@test substitute(ex, Dict(τ => 0.5)) == cos(0.5) * A(0.5) # true","category":"page"},{"location":"symbolics/#Symbolic-Derivatives","page":"Using with Symbolics/ModelingToolkit","title":"Symbolic Derivatives","text":"","category":"section"},{"location":"symbolics/","page":"Using with Symbolics/ModelingToolkit","title":"Using with Symbolics/ModelingToolkit","text":"D = Differential(τ)\n\nex1 = A(τ)\n\n# Derivative of interpolation\nex2 = expand_derivatives(D(ex1))\n\n@test substitute(ex2, Dict(τ => 0.5)) == DataInterpolations.derivative(A, 0.5) # true\n\n# Higher Order Derivatives\nex3 = expand_derivatives(D(D(A(τ))))\n\n@test substitute(ex3, Dict(τ => 0.5)) == DataInterpolations.derivative(A, 0.5, 2) # true","category":"page"},{"location":"symbolics/#Using-with-ModelingToolkit.jl","page":"Using with Symbolics/ModelingToolkit","title":"Using with ModelingToolkit.jl","text":"","category":"section"},{"location":"symbolics/","page":"Using with Symbolics/ModelingToolkit","title":"Using with Symbolics/ModelingToolkit","text":"Most common use case with ModelingToolkit.jl is to plug in interpolation objects as input functions. This can be done using TimeVaryingFunction component of ModelingToolkitStandardLibrary.jl.","category":"page"},{"location":"symbolics/","page":"Using with Symbolics/ModelingToolkit","title":"Using with Symbolics/ModelingToolkit","text":"using DataInterpolations\nusing ModelingToolkitStandardLibrary.Blocks\nusing ModelingToolkit\nusing ModelingToolkit: t_nounits as t, D_nounits as D\nusing OrdinaryDiffEq\n\nus = [0.0, 1.5, 0.0]\ntimes = [0.0, 0.5, 1.0]\nA = LinearInterpolation(us, times)\n\n@named src = TimeVaryingFunction(A)\nvars = @variables x(t) out(t)\neqs = [out ~ src.output.u, D(x) ~ 1 + out]\n@named sys = ODESystem(eqs, t, vars, []; systems = [src])\n\nsys = structural_simplify(sys)\npro
[…]
