+++ versions.js
@@ -1,5 +1,6 @@
 var DOC_VERSIONS = [
   "stable",
+  "v6.1",
   "v6.0",
   "v5.4",
   "v5.3",
@@ -15,5 +16,5 @@ var DOC_VERSIONS = [
   "v4.1",
   "dev",
 ];
-var DOCUMENTER_NEWEST = "v6.0.0";
+var DOCUMENTER_NEWEST = "v6.1.0";
 var DOCUMENTER_STABLE = "stable";
