### 2024-08-01 c9a1e94
Fix problem in get_idx when f = +/-Inf
