+++ src/interpolation_caches.jl
@@ -25,13 +25,13 @@ struct LinearInterpolation{uType, tType, IType, pType, T} <: AbstractInterpolati
     I::IType
     p::LinearParameterCache{pType}
     extrapolate::Bool
-    idx_prev::Base.RefValue{Int}
+    iguesser::Guesser{tType}
     cache_parameters::Bool
     linear_lookup::Bool
     function LinearInterpolation(u, t, I, p, extrapolate, cache_parameters, assume_linear_t)
         linear_lookup = seems_linear(assume_linear_t, t)
         new{typeof(u), typeof(t), typeof(I), typeof(p.slope), eltype(u)}(
-            u, t, I, p, extrapolate, Ref(1), cache_parameters, linear_lookup)
+            u, t, I, p, extrapolate, Guesser(t), cache_parameters, linear_lookup)
     end
 end
 
@@ -73,7 +73,7 @@ struct QuadraticInterpolation{uType, tType, IType, pType, T} <: AbstractInterpol
     p::QuadraticParameterCache{pType}
     mode::Symbol
     extrapolate::Bool
-    idx_prev::Base.RefValue{Int}
+    iguesser::Guesser{tType}
     cache_parameters::Bool
     linear_lookup::Bool
     function QuadraticInterpolation(
@@ -82,7 +82,7 @@ struct QuadraticInterpolation{uType, tType, IType, pType, T} <: AbstractInterpol
             error("mode should be :Forward or :Backward for QuadraticInterpolation")
         linear_lookup = seems_linear(assume_linear_t, t)
         new{typeof(u), typeof(t), typeof(I), typeof(p.l₀), eltype(u)}(
-            u, t, I, p, mode, extrapolate, Ref(1), cache_parameters, linear_lookup)
+            u, t, I, p, mode, extrapolate, Guesser(t), cache_parameters, linear_lookup)
     end
 end
 
@@ -124,7 +124,7 @@ struct LagrangeInterpolation{uType, tType, T, bcacheType} <:
     bcache::bcacheType
     idxs::Vector{Int}
     extrapolate::Bool
-    idx_prev::Base.RefValue{Int}
+    iguesser::Guesser{tType}
     function LagrangeInterpolation(u, t, n, extrapolate)
         bcache = zeros(eltype(u[1]), n + 1)
         idxs = zeros(Int, n + 1)
@@ -135,7 +135,7 @@ struct LagrangeInterpolation{uType, tType, T, bcacheType} <:
             bcache,
             idxs,
             extrapolate,
-            Ref(1)
+            Guesser(t)
         )
     end
 end
@@ -178,7 +178,7 @@ struct AkimaInterpolation{uType, tType, IType, bType, cType, dType, T} <:
     c::cType
     d::dType
     extrapolate::Bool
-    idx_prev::Base.RefValue{Int}
+    iguesser::Guesser{tType}
     cache_parameters::Bool
     linear_lookup::Bool
     function AkimaInterpolation(
@@ -192,7 +192,7 @@ struct AkimaInterpolation{uType, tType, IType, bType, cType, dType, T} <:
             c,
             d,
             extrapolate,
-            Ref(1),
+            Guesser(t),
             cache_parameters,
             linear_lookup
         )
@@ -258,14 +258,14 @@ struct ConstantInterpolation{uType, tType, IType, T} <: AbstractInterpolation{T}
     p::Nothing
     dir::Symbol # indicates if value to the $dir should be used for the interpolation
     extrapolate::Bool
-    idx_prev::Base.RefValue{Int}
+    iguesser::Guesser{tType}
     cache_parameters::Bool
     linear_lookup::Bool
     function ConstantInterpolation(
[…]
