### 2024-08-02 6b25a7b
Stop having _interpolate and _derivative return the index
