+++ src/derivatives.jl
@@ -1,11 +1,11 @@
 function derivative(A, t, order = 1; iguess = A.iguesser(t))
     ((t < A.t[1] || t > A.t[end]) && !A.extrapolate) && throw(ExtrapolationError())
-    
+
     return if order == 1
-        _derivative(A, t, iguess)[1]
+        _derivative(A, t, iguess)
     elseif order == 2
         ForwardDiff.derivative(t -> begin
-                _derivative(A, t, iguess)[1]
+                _derivative(A, t, iguess)
             end, t)
     else
         throw(DerivativeNotFoundError())
@@ -24,7 +24,7 @@ function _derivative(A::QuadraticInterpolation, t::Number, iguess)
     du₀ = l₀ * (2t - A.t[i₁] - A.t[i₂])
     du₁ = l₁ * (2t - A.t[i₀] - A.t[i₂])
     du₂ = l₂ * (2t - A.t[i₀] - A.t[i₁])
-    return @views @. du₀ + du₁ + du₂, i₀
+    return @views @. du₀ + du₁ + du₂
 end
 
 function _derivative(A::LagrangeInterpolation{<:AbstractVector}, t::Number)
@@ -96,21 +96,21 @@ function _derivative(A::LagrangeInterpolation{<:AbstractMatrix}, t::Number)
 end
 
 function _derivative(A::LagrangeInterpolation{<:AbstractVector}, t::Number, idx)
-    _derivative(A, t), idx
+    _derivative(A, t)
 end
 function _derivative(A::LagrangeInterpolation{<:AbstractMatrix}, t::Number, idx)
-    _derivative(A, t), idx
+    _derivative(A, t)
 end
 
 function _derivative(A::AkimaInterpolation{<:AbstractVector}, t::Number, iguess)
     idx = get_idx(A, t, iguess; idx_shift = -1, side = :first)
     j = min(idx, length(A.c))  # for smooth derivative at A.t[end]
     wj = t - A.t[idx]
-    (@evalpoly wj A.b[idx] 2A.c[j] 3A.d[j]), idx
+    @evalpoly wj A.b[idx] 2A.c[j] 3A.d[j]
 end
 
 function _derivative(A::ConstantInterpolation, t::Number, iguess)
-    return zero(first(A.u)), iguess
+    return zero(first(A.u))
 end
 
 function _derivative(A::ConstantInterpolation{<:AbstractVector}, t::Number)
@@ -127,7 +127,7 @@ end
 function _derivative(A::QuadraticSpline{<:AbstractVector}, t::Number, iguess)
     idx = get_idx(A, t, iguess; lb = 2, ub_shift = 0, side = :first)
     σ = get_parameters(A, idx - 1)
-    A.z[idx - 1] + 2σ * (t - A.t[idx - 1]), idx
+    A.z[idx - 1] + 2σ * (t - A.t[idx - 1])
 end
 
 # CubicSpline Interpolation
@@ -139,7 +139,7 @@ function _derivative(A::CubicSpline{<:AbstractVector}, t::Number, iguess)
     c₁, c₂ = get_parameters(A, idx)
     dC = c₁
     dD = -c₂
-    dI + dC + dD, idx
+    dI + dC + dD
 end
 
 function _derivative(A::BSplineInterpolation{<:AbstractVector{<:Number}}, t::Number, iguess)
@@ -160,7 +160,7 @@ function _derivative(A::BSplineInterpolation{<:AbstractVector{<:Number}}, t::Num
[…]
