+++ src/interpolation_methods.jl
@@ -1,7 +1,7 @@
 function _interpolate(A, t; iguess = A.iguesser(t))
     ((t < A.t[1] || t > A.t[end]) && !A.extrapolate) &&
         throw(ExtrapolationError())
-    return _interpolate(A, t, iguess)[1]
+    return _interpolate(A, t, iguess)
 end
 
 # Linear Interpolation
@@ -30,14 +30,14 @@ function _interpolate(A::LinearInterpolation{<:AbstractVector}, t::Number, igues
     end
     val = oftype(Δu, val)
 
-    val, idx
+    val
 end
 
 function _interpolate(A::LinearInterpolation{<:AbstractMatrix}, t::Number, iguess)
     idx = get_idx(A, t, iguess)
     Δt = t - A.t[idx]
     slope = get_parameters(A, idx)
-    return A.u[:, idx] + slope * Δt, idx
+    return A.u[:, idx] + slope * Δt
 end
 
 # Quadratic Interpolation
@@ -53,7 +53,7 @@ function _interpolate(A::QuadraticInterpolation, t::Number, iguess)
     u₀ = l₀ * (t - A.t[i₁]) * (t - A.t[i₂])
     u₁ = l₁ * (t - A.t[i₀]) * (t - A.t[i₂])
     u₂ = l₂ * (t - A.t[i₀]) * (t - A.t[i₁])
-    return u₀ + u₁ + u₂, i₀
+    return u₀ + u₁ + u₂
 end
 
 # Lagrange Interpolation
@@ -61,7 +61,7 @@ function _interpolate(A::LagrangeInterpolation{<:AbstractVector}, t::Number, igu
     idx = get_idx(A, t, iguess)
     findRequiredIdxs!(A, t, idx)
     if A.t[A.idxs[1]] == t
-        return A.u[A.idxs[1]], idx
+        return A.u[A.idxs[1]]
     end
     N = zero(A.u[1])
     D = zero(A.t[1])
@@ -83,14 +83,14 @@ function _interpolate(A::LagrangeInterpolation{<:AbstractVector}, t::Number, igu
         D += tmp
         N += (tmp * A.u[A.idxs[i]])
     end
-    N / D, idx
+    N / D
 end
 
 function _interpolate(A::LagrangeInterpolation{<:AbstractMatrix}, t::Number, iguess)
     idx = get_idx(A, t, iguess)
     findRequiredIdxs!(A, t, idx)
     if A.t[A.idxs[1]] == t
-        return A.u[:, A.idxs[1]], idx
+        return A.u[:, A.idxs[1]]
     end
     N = zero(A.u[:, 1])
     D = zero(A.t[1])
@@ -112,13 +112,13 @@ function _interpolate(A::LagrangeInterpolation{<:AbstractMatrix}, t::Number, igu
         D += tmp
         @. N += (tmp * A.u[:, A.idxs[i]])
     end
-    N / D, idx
+    N / D
 end
 
 function _interpolate(A::AkimaInterpolation{<:AbstractVector}, t::Number, iguess)
     idx = get_idx(A, t, iguess)
     wj = t - A.t[idx]
-    (@evalpoly wj A.u[idx] A.b[idx] A.c[idx] A.d[idx]), idx
+    @evalpoly wj A.u[idx] A.b[idx] A.c[idx] A.d[idx]
 end
 
 # ConstantInterpolation Interpolation
@@ -130,7 +130,7 @@ function _interpolate(A::ConstantInterpolation{<:AbstractVector}, t::Number, igu
         # :right means that value to the right is used for interpolation
[…]
