+++ test/derivative_tests.jl
@@ -39,7 +39,7 @@ function test_derivatives(method; args = [], kwargs = [], name::String)
             @test isapprox(fdiff, adiff, atol = 1e-8)
             @test isapprox(fdiff2, adiff2, atol = 1e-8)
             # Cached index
-            if hasproperty(func, :iguesser)
+            if hasproperty(func, :iguesser) && !func.iguesser.linear_lookup
                 @test abs(func.iguesser.idx_prev[] -
                           searchsortedfirstcorrelated(func.t, _t, func.iguesser(_t))) <= 1
             end
