### 2024-08-06 1a98506
POC with ConstantInterpolation
