### 2024-08-08 #6917 @dawbarton
Sorry for the delay with this. Overall, it looks good to me, particularly with carstenbauer's comments.

A few minor comments on the paper:

1. Missing a space between the first and second sentences of the Summary.
2. Missing a reference for Symbolics.jl in the Summary.
3. First sentence of the Example could do with rewording. Maybe "The following tutorials are provided in the documentation. Tutorial 1 provides.... Tutorial 2 provides.... Tutorial 3 provides...".

As a side note, it might be worth a mention somewhere in the documentation that vectors-of-vectors can be used for the input to get multiple outputs from DataInterpolations. (This wasn't obvious to me and it doesn't accept a matrix as an input.)
