### 2024-08-11 596dfe8
docs: fix references, spacing and language in the example section in joss paper
