+++ dev/assets/Project.toml
@@ -20,4 +20,4 @@ OrdinaryDiffEq = "6"
 Plots = "1"
 RegularizationTools = "0.6"
 StableRNGs = "1"
-Symbolics = "5.29"
+Symbolics = "5.29, 6.0"
