### 2024-08-12 #6917 @sathvikbhagavan
Thanks @dawbarton for the review. I have addressed all the three points in the paper. Do take a look if any other further changes are needed.

For interpolations with vector of vectors and matrices, I haven't mentioned this in the documentation because it is not fully supported yet (Not all methods support it). There is an open issue about it - https://github.com/SciML/DataInterpolations.jl/issues/206. I plan to tackle it soon.
