+++ Project.toml
@@ -39,7 +39,7 @@ Reexport = "1"
 RegularizationTools = "0.6"
 SafeTestsets = "0.1"
 StableRNGs = "1"
-Symbolics = "5.29"
+Symbolics = "5.29, 6"
 Test = "1"
 Zygote = "0.6.70"
 julia = "1.10"
