+++ versions.js
@@ -1,5 +1,6 @@
 var DOC_VERSIONS = [
   "stable",
+  "v6.3",
   "v6.2",
   "v6.1",
   "v6.0",
@@ -17,5 +18,5 @@ var DOC_VERSIONS = [
   "v4.1",
   "dev",
 ];
-var DOCUMENTER_NEWEST = "v6.2.0";
+var DOCUMENTER_NEWEST = "v6.3.0";
 var DOCUMENTER_STABLE = "stable";
