### 2024-08-28 dec4c30
docs: change the order of authors in joss paper
