+++ versions.js
@@ -1,5 +1,6 @@
 var DOC_VERSIONS = [
   "stable",
+  "v6.4",
   "v6.3",
   "v6.2",
   "v6.1",
@@ -18,5 +19,5 @@ var DOC_VERSIONS = [
   "v4.1",
   "dev",
 ];
-var DOCUMENTER_NEWEST = "v6.3.0";
+var DOCUMENTER_NEWEST = "v6.4.0";
 var DOCUMENTER_STABLE = "stable";
