### 2024-09-02 a85218a
build(deps): bump julia-actions/setup-julia from 2.3.0 to 2.4.0
Bumps [julia-actions/setup-julia](https://github.com/julia-actions/setup-julia) from 2.3.0 to 2.4.0.
- [Release notes](https://github.com/julia-actions/setup-julia/releases)
- [Commits](https://github.com/julia-actions/setup-julia/compare/v2.3...v2.4)

---
updated-dependencies:
- dependency-name: julia-actions/setup-julia
  dependency-type: direct:production
  update-type: version-update:semver-minor
...

Signed-off-by: dependabot[bot] <support@github.com>
