+++ .github/workflows/CompatHelper.yml
@@ -13,7 +13,7 @@ jobs:
   CompatHelper:
     runs-on: ubuntu-latest
     steps:
-      - uses: julia-actions/setup-julia@780022b48dfc0c2c6b94cfee6a9284850107d037
+      - uses: julia-actions/setup-julia@512e2ab44f5848ea62bddd21d1ee906e989b0360
         with:
           version: 1.3
       - name: Pkg.add("CompatHelper")
