### 2024-09-04 #6917 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
